### 2018-04-25 #703 @arfon
@rcurtin - many thanks for your submission. Is there anything stopping you from moving the paper into the https://github.com/mlpack/mlpack repository? This is our preferred way of handling submissions.

The `paper.md` file can exist basically anywhere in the https://github.com/mlpack/mlpack/ repo (e.g. nested somewhere under the docs folder)
