### 2018-04-25 #703 @rcurtin
@arfon thanks for the quick reply.  If it's necessary, we can move the paper into the repository, but to me it seems that the paper itself is a separate artifact from the code.  (As time goes on, the paper itself may become out of date, for example.)  I think that the Armadillo submission, which I was also involved with, had the paper in a separate place also.  Let me know what you think.

I see that there are some minor issues with the paper; I will handle those.  But I do have a question: some of our authors chose to have no affiliation for this paper, but it looks like the paper has compiled such that their affiliation has default-chosen the first affiliation.  Is there a way to specify no affiliation?  I don't see anything in the documentation.
