### 2018-04-25 8908438
Avoid matrix copy in KDE main
