### 2018-04-26 df0b505
Cover edge case to suppress compiler warning
The variable 'model' may not be initialized, thus causing a compile warning in
src/mlpack/methods/hoeffding_trees/hoeffding_tree_main.cpp:251
