### 2018-04-26 #703 @arfon
> @arfon thanks for the quick reply. If it's necessary, we can move the paper into the repository, but to me it seems that the paper itself is a separate artifact from the code. (As time goes on, the paper itself may become out of date, for example.) I think that the Armadillo submission, which I was also involved with, had the paper in a separate place also. Let me know what you think.

This is a hard requirement for us these days. Our editorial system is heavily automated and so the current layout (paper not being with the software) means manual work for our editorial team.
