### 2018-04-26 6302607
Delete unused variable
