### 2018-04-26 2e4cdd1
add simple code to statisfy the independence and SmallMvuSdp works
