+++ src/mlpack/tests/sdp_primal_dual_test.cpp
@@ -1,6 +1,7 @@
 /**
  * @file sdp_primal_dual_test.cpp
  * @author Stephen Tu
+ * @author Kaiqiang Xu
  *
  *
  * mlpack is free software; you may redistribute it and/or modify it under the
@@ -13,6 +14,7 @@
 #include <mlpack/core/optimizers/sdp/primal_dual.hpp>
 #include <mlpack/methods/neighbor_search/neighbor_search.hpp>
 
+//#include <set>
 #include <boost/test/unit_test.hpp>
 #include "test_tools.hpp"
 
@@ -557,92 +559,137 @@ BOOST_AUTO_TEST_CASE(CorrelationCoeffToySdp)
   BOOST_REQUIRE_CLOSE(obj, 2 * (-0.978), 1e-3);
 }
 
-// /**
-// * Maximum variance unfolding (MVU) SDP to learn the unrolled gram matrix. For
-// * the SDP formulation, see:
-// *
-// *   Unsupervised learning of image manifolds by semidefinite programming.
-// *   Kilian Weinberger and Lawrence Saul. CVPR 04.
-// *   http://repository.upenn.edu/cgi/viewcontent.cgi?article=1000&context=cis_papers
-// *
-// * @param origData origDim x numPoints
-// * @param numNeighbors
-// */
-// static inline SDP<arma::sp_mat> ConstructMvuSDP(const arma::mat& origData,
-//                                                size_t numNeighbors)
-// {
-//  const size_t numPoints = origData.n_cols;
+ /**
+ * Maximum variance unfolding (MVU) SDP to learn the unrolled gram matrix. For
+ * the SDP formulation, see:
+ *
+ *   Unsupervised learning of image manifolds by semidefinite programming.
+ *   Kilian Weinberger and Lawrence Saul. CVPR 04.
+ *   http://repository.upenn.edu/cgi/viewcontent.cgi?article=1000&context=cis_papers
+ *
+ * Here is another paper which clarifies more specific SDP formulation of MVU problem.
+ *
+ *   Scalable semidefinite manifold learning
+ *   Nikolaos Vasiloglou, Alexander G. Gray, David V. Anderson.
+ *   2008 IEEE Workshop on Machine Learning for Signal Processing
+ *
+ * @param origData origDim x numPoints
+ * @param numNeighbors
+ */
+ static inline SDP<arma::sp_mat> ConstructMvuSDP(const arma::mat& origData,
+                                                size_t numNeighbors)
+ {
+  const size_t numPoints = origData.n_cols;
 
-//  assert(numNeighbors <= numPoints);
+  assert(numNeighbors <= numPoints);
 
-//  arma::Mat<size_t> neighbors;
-//  arma::mat distances;
-//  KNN knn(origData);
-//  knn.Search(numNeighbors, neighbors, distances);
+  arma::Mat<size_t> neighbors;
+  arma::mat distances;
+  KNN knn(origData);
+  knn.Search(numNeighbors, neighbors, distances);
+  // how to construct the matrices A, b, and C? Plz refer to the above paper.
+  std::set<std::pair<int, int>> neighbors_set;
+  for (size_t i = 0; i < neighbors.n_cols; ++i)
+  {
+     for (size_t j = 0; j < numNeighbors; ++j)
+     {
+       size_t x = i, y = neighbors(j, i);
+       if (x > y) {
+           x = neighbors(j, i);
+           y = i;
+       }
+       if (neighbors_set.find(std::make_pair(x, y)) != neighbors_set.end()) {
+           continue;
[…]
