### 2018-04-26 164c22d
Delete leafSize parameter for KDE trees
A new constructor with a tree as a parameter
will handle different leaf sizes
