### 2018-04-26 b776959
Editing core.hpp & COPYRIGHT.txt
