### 2018-04-30 ab00121
Add LSH and Radical bindings tests
