+++ src/mlpack/tests/main_tests/lsh_test.cpp
@@ -1,0 +1,357 @@
+/**
+ * @file lsh_test.cpp
+ * @author Manish Kumar
+ *
+ * Test mlpackMain() of lsh_main.cpp.
+ *
+ * mlpack is free software; you may redistribute it and/or modify it under the
+ * terms of the 3-clause BSD license.  You should have received a copy of the
+ * 3-clause BSD license along with mlpack.  If not, see
+ * http://www.opensource.org/licenses/BSD-3-Clause for more information.
+ */
+#include <string>
+
+#define BINDING_TYPE BINDING_TYPE_TEST
+static const std::string testName = "LSH";
+
+#include <mlpack/core.hpp>
+#include <mlpack/core/util/mlpack_main.hpp>
+#include "test_helper.hpp"
+#include <mlpack/methods/lsh/lsh_main.cpp>
+
+#include <boost/test/unit_test.hpp>
+#include "../test_tools.hpp"
+
+using namespace mlpack;
+
+struct LSHTestFixture
+{
+ public:
+  LSHTestFixture()
+  {
+    // Cache in the options for this program.
+    CLI::RestoreSettings(testName);
+  }
+
+  ~LSHTestFixture()
+  {
+    // Clear the settings.
+    bindings::tests::CleanMemory();
+    CLI::ClearSettings();
+  }
+};
+
+BOOST_FIXTURE_TEST_SUITE(LSHMainTest, LSHTestFixture);
+
+/**
+ * Check that output neighbors and distances have valid dimensions.
+ */
+BOOST_AUTO_TEST_CASE(LSHOutputDimensionTest)
+{
+  arma::mat reference = arma::randu<arma::mat>(5, 100);
+
+  SetInputParam("reference", std::move(reference));
+  SetInputParam("k", (int) 6);
+
+  mlpackMain();
+
+  // Check the neighbors matrix has 6 points for each of the 100 input points.
+  BOOST_REQUIRE_EQUAL(CLI::GetParam<arma::Mat<size_t>>("neighbors").n_rows, 6);
+  BOOST_REQUIRE_EQUAL(CLI::GetParam<arma::Mat<size_t>>("neighbors").n_cols, 100);
+
+  // Check the distances matrix has 6 points for each of the 100 input points.
+  BOOST_REQUIRE_EQUAL(CLI::GetParam<arma::mat>("distances").n_rows, 6);
+  BOOST_REQUIRE_EQUAL(CLI::GetParam<arma::mat>("distances").n_cols, 100);
+}
+
+/**
+ * Ensure that bucket_size, second_hash_size & number of nearest neighbors
+ * are always positive.
+ */
+BOOST_AUTO_TEST_CASE(LSHParamValidityTest)
+{
+  arma::mat reference = arma::randu<arma::mat>(5, 100);
+
+  // Test for bucket_size.
+
+  SetInputParam("reference", reference);
+  SetInputParam("k", (int) 6);
+  SetInputParam("bucket_size", (int) -1.0);
+
+  Log::Fatal.ignoreInput = true;
+  BOOST_REQUIRE_THROW(mlpackMain(), std::runtime_error);
+  Log::Fatal.ignoreInput = false;
+
+  bindings::tests::CleanMemory();
+
+  // Test for bucket_size.
+
+  SetInputParam("reference", reference);
+  SetInputParam("k", (int) 6);
+  SetInputParam("second_hash_size", (int) -1.0);
+
+  Log::Fatal.ignoreInput = true;
+  BOOST_REQUIRE_THROW(mlpackMain(), std::runtime_error);
+  Log::Fatal.ignoreInput = false;
+
+  bindings::tests::CleanMemory();
+
+  // Test for number of nearest neighbors.
+
+  SetInputParam("reference", std::move(reference));
[…]
