### 2018-04-30 a28143b
Minor style fixes (cite style, indentation, comment style).
