### 2018-04-30 #703 @rcurtin
Ok, if it is a hard requirement, we have gone ahead and merged the paper.md into the mlpack/mlpack repository.  The repository is now:

https://github.com/mlpack/mlpack

and the paper is located at:

https://github.com/mlpack/mlpack/tree/master/doc/joss_paper

Please let me know if that works, and if I didn't do it right just let me know and I can fix it.  Thanks!
