+++ src/mlpack/methods/ann/layer/transposed_convolution.hpp
@@ -1,0 +1,345 @@
+/**
+ * @file transposed_convolution.hpp
+ * @author Marcus Edel
+ * @author Shikhar Jaiswal
+ *
+ * Definition of the Transposed Convolution module class.
+ *
+ * mlpack is free software; you may redistribute it and/or modify it under the
+ * terms of the 3-clause BSD license.  You should have received a copy of the
+ * 3-clause BSD license along with mlpack.  If not, see
+ * http://www.opensource.org/licenses/BSD-3-Clause for more information.
+ */
+#ifndef MLPACK_METHODS_ANN_LAYER_TRANSPOSED_CONVOLUTION_HPP
+#define MLPACK_METHODS_ANN_LAYER_TRANSPOSED_CONVOLUTION_HPP
+
+#include <mlpack/prereqs.hpp>
+
+#include <mlpack/methods/ann/convolution_rules/border_modes.hpp>
+#include <mlpack/methods/ann/convolution_rules/naive_convolution.hpp>
+#include <mlpack/methods/ann/convolution_rules/fft_convolution.hpp>
+#include <mlpack/methods/ann/convolution_rules/svd_convolution.hpp>
+
+#include "layer_types.hpp"
+
+namespace mlpack {
+namespace ann /** Artificial Neural Network. */ {
+
+/**
+ * Implementation of the Transposed Convolution class. The Transposed 
+ * Convolution class represents a single layer of a neural network.
+ *
+ * @tparam ForwardConvolutionRule Convolution to perform forward process.
+ * @tparam BackwardConvolutionRule Convolution to perform backward process.
+ * @tparam GradientConvolutionRule Convolution to calculate gradient.
+ * @tparam InputDataType Type of the input data (arma::colvec, arma::mat,
+ *         arma::sp_mat or arma::cube).
+ * @tparam OutputDataType Type of the output data (arma::colvec, arma::mat,
+ *         arma::sp_mat or arma::cube).
+ */
+template <
+    typename ForwardConvolutionRule = NaiveConvolution<ValidConvolution>,
+    typename BackwardConvolutionRule = NaiveConvolution<FullConvolution>,
+    typename GradientConvolutionRule = NaiveConvolution<ValidConvolution>,
+    typename InputDataType = arma::mat,
+    typename OutputDataType = arma::mat
+>
+class TransposedConvolution
+{
+ public:
+  //! Create the Transposed Convolution object.
+  TransposedConvolution();
+
+  /**
+   * Create the Transposed Convolution object using the specified number of input maps,
+   * output maps, filter size, stride and padding parameter.
+   *
+   * @param inSize The number of input maps.
+   * @param outSize The number of output maps.
+   * @param kW Width of the filter/kernel.
+   * @param kH Height of the filter/kernel.
+   * @param dW Stride of filter application in the x direction.
+   * @param dH Stride of filter application in the y direction.
+   * @param padW Padding width of the input.
+   * @param padH Padding height of the input.
+   * @param inputWidth The widht of the input data.
+   * @param inputHeight The height of the input data.
+   */
+  TransposedConvolution(const size_t inSize,
+                        const size_t outSize,
+                        const size_t kW,
+                        const size_t kH,
+                        const size_t dW = 1,
+                        const size_t dH = 1,
+                        const size_t padW = 0,
+                        const size_t padH = 0,
+                        const size_t inputWidth = 0,
+                        const size_t inputHeight = 0);
+
+  /*
+   * Set the weight and bias term.
+   */
+  void Reset();
+
+  /**
+   * Ordinary feed forward pass of a neural network, evaluating the function
[…]
