+++ src/mlpack/methods/ann/layer/layer.hpp
@@ -29,5 +29,6 @@
 #include "sequential.hpp"
 #include "concat.hpp"
 #include "vr_class_reward.hpp"
+#include "transposed_convolution.hpp"
 
 #endif
