### 2018-05-01 68f7762
Make sure we have all the methods that we need.
