### 2018-05-01 9ae8a89
Get the first objective at the first iteration.
