+++ src/mlpack/tests/main_tests/lsh_test.cpp
@@ -57,7 +57,8 @@ BOOST_AUTO_TEST_CASE(LSHOutputDimensionTest)
 
   // Check the neighbors matrix has 6 points for each of the 100 input points.
   BOOST_REQUIRE_EQUAL(CLI::GetParam<arma::Mat<size_t>>("neighbors").n_rows, 6);
-  BOOST_REQUIRE_EQUAL(CLI::GetParam<arma::Mat<size_t>>("neighbors").n_cols, 100);
+  BOOST_REQUIRE_EQUAL(CLI::GetParam<arma::Mat<size_t>>("neighbors").n_cols,
+                      100);
 
   // Check the distances matrix has 6 points for each of the 100 input points.
   BOOST_REQUIRE_EQUAL(CLI::GetParam<arma::mat>("distances").n_rows, 6);
@@ -84,7 +85,7 @@ BOOST_AUTO_TEST_CASE(LSHParamValidityTest)
 
   bindings::tests::CleanMemory();
 
-  // Test for bucket_size.
+  // Test for second_hash_size.
 
   SetInputParam("reference", reference);
   SetInputParam("k", (int) 6);
@@ -135,6 +136,7 @@ BOOST_AUTO_TEST_CASE(LSHDiffTablesTest)
   SetInputParam("reference", reference);
   SetInputParam("k", (int) 6);
 
+  mlpack::math::FixedRandomSeed();
   mlpackMain();
 
   arma::Mat<size_t> neighbors = CLI::GetParam<arma::Mat<size_t>>("neighbors");
@@ -148,6 +150,7 @@ BOOST_AUTO_TEST_CASE(LSHDiffTablesTest)
   SetInputParam("k", (int) 6);
   SetInputParam("tables", (int) 40);
 
+  mlpack::math::FixedRandomSeed();
   mlpackMain();
 
   // Check that initial outputs and final outputs using two models are
@@ -168,6 +171,7 @@ BOOST_AUTO_TEST_CASE(LSHDiffProjectionsTest)
   SetInputParam("reference", reference);
   SetInputParam("k", (int) 6);
 
+  mlpack::math::FixedRandomSeed();
   mlpackMain();
 
   arma::Mat<size_t> neighbors = CLI::GetParam<arma::Mat<size_t>>("neighbors");
@@ -181,6 +185,7 @@ BOOST_AUTO_TEST_CASE(LSHDiffProjectionsTest)
   SetInputParam("k", (int) 6);
   SetInputParam("projections", (int) 30);
 
+  mlpack::math::FixedRandomSeed();
   mlpackMain();
 
   // Check that initial outputs and final outputs using two models are
@@ -201,6 +206,7 @@ BOOST_AUTO_TEST_CASE(LSHDiffHashWidthTest)
   SetInputParam("reference", reference);
   SetInputParam("k", (int) 6);
 
+  mlpack::math::FixedRandomSeed();
   mlpackMain();
 
   arma::Mat<size_t> neighbors = CLI::GetParam<arma::Mat<size_t>>("neighbors");
@@ -214,6 +220,7 @@ BOOST_AUTO_TEST_CASE(LSHDiffHashWidthTest)
   SetInputParam("k", (int) 6);
   SetInputParam("hash_width", (double) 0.5);
 
+  mlpack::math::FixedRandomSeed();
   mlpackMain();
 
   // Check that initial outputs and final outputs using two models are
@@ -230,8 +237,10 @@ BOOST_AUTO_TEST_CASE(LSHDiffHashWidthTest)
 BOOST_AUTO_TEST_CASE(LSHDiffNumProbesTest)
 {
   arma::mat reference = arma::randu<arma::mat>(5, 100);
+  arma::mat query = arma::randu<arma::mat>(5, 40);
 
[…]
