+++ src/mlpack/methods/lsh/lsh_main.cpp
@@ -201,6 +201,12 @@ static void mlpackMain()
     // Load the true neighbors.
     arma::Mat<size_t> trueNeighbors =
         std::move(CLI::GetParam<arma::Mat<size_t>>("true_neighbors"));
+
+    if (trueNeighbors.n_rows != neighbors.n_rows ||
+          trueNeighbors.n_cols != neighbors.n_cols)
+        Log::Fatal << "The true neighbors file must have the same number of "
+            << "values as the set of neighbors being queried!" << endl;
+
     Log::Info << "Using true neighbor indices from '"
         << CLI::GetPrintableParam<arma::Mat<size_t>>("true_neighbors") << "'."
         << endl;
