### 2018-05-02 6137b32
fix naming conventions and some comments^Cf code
