### 2018-05-02 ba9f83d
Improve KDE API
