### 2018-05-03 c160eef
set random seed and change pair to tuple
