+++ src/mlpack/tests/sdp_primal_dual_test.cpp
@@ -587,8 +587,15 @@ static inline SDP<arma::sp_mat> ConstructMvuSDP(const arma::mat& origData,
   KNN knn(origData);
   knn.Search(numNeighbors, neighbors, distances);
 
-  // how to construct the matrices A, b, and C? Plz refer to the above paper.
-  std::set<std::pair<int, int>> neighborsSet;
+  typedef std::tuple<size_t, size_t, double_t> neighborTupleType;
+//  auto comp = [](const neighborTupleType& t1,
+//                 const neighborTupleType& t2)
+//  {
+//    return std::get<0>(t1) != std::get<0>(t2) or
+//            std::get<1>(t1) != std::get<1>(t2);
+//  };
+//  auto neighborsSet = std::set<neighborTupleType, decltype(comp)>(comp);
+  std::set<neighborTupleType > neighborsSet;
   for (size_t i = 0; i < neighbors.n_cols; ++i)
   {
     for (size_t j = 0; j < numNeighbors; ++j)
@@ -599,66 +606,44 @@ static inline SDP<arma::sp_mat> ConstructMvuSDP(const arma::mat& origData,
         x = neighbors(j, i);
         y = i;
       }
-
-      if (neighborsSet.find(std::make_pair(x, y)) != neighborsSet.end())
-      {
-        continue;
-      }
-
-      neighborsSet.insert(std::make_pair(x, y));
+      neighborsSet.insert(std::make_tuple(x, y, distances(j, i)));
     }
   }
-
+  // how to construct the matrices A, b, and C? Plz refer to the above paper.
   SDP<arma::sp_mat> sdp(numPoints, neighborsSet.size(), 1);
   sdp.C().eye(numPoints, numPoints);
   sdp.C() *= -1;
   sdp.DenseA()[0].ones(numPoints, numPoints);
   sdp.DenseB()[0] = 0;
 
-  neighborsSet.clear();
   size_t index = 0;
 
-  for (size_t i = 0; i < neighbors.n_cols; ++i)
+  for (auto it = neighborsSet.begin(); it != neighborsSet.end(); it++)
   {
-    for (size_t j = 0; j < numNeighbors; ++j)
-    {
-      // The set is used to filter duplicate neighbor pair.
-      // Then the linear indenpence of the constraints can be satisfied.
-      size_t x = i, y = neighbors(j, i);
-      if (x > y)
-      {
-        x = neighbors(j, i);
-        y = i;
-      }
-      if (neighborsSet.find(std::make_pair(x, y)) != neighborsSet.end())
-      {
-        continue;
-      }
-      neighborsSet.insert(std::make_pair(x, y));
+    size_t x = std::get<0>(*it);
+    size_t y = std::get<1>(*it);
+    double_t d = std::get<2>(*it);
 
-      // This is the index of the constraint.
-      arma::sp_mat& aRef = sdp.SparseA()[index];
-      aRef.zeros(numPoints, numPoints);
+    arma::sp_mat& aRef = sdp.SparseA()[index];
+    aRef.zeros(numPoints, numPoints);
 
-      // A_ij(i, i) = 1.
-      aRef(i, i) = 1;
+    // A_ij(i, i) = 1.
+    aRef(x, x) = 1;
 
-      // A_ij(i, j) = -1.
-      aRef(i, neighbors(j, i)) = -1;
+    // A_ij(i, j) = -1.
+    aRef(x, y) = -1;
 
-      // A_ij(j, i) = -1.
-      aRef(neighbors(j, i), i) = -1;
+    // A_ij(j, i) = -1.
+    aRef(y, x) = -1;
 
[…]
