### 2018-05-03 5309e42
Added tests to radical
