### 2018-05-03 ead87a5
Handle FirstPointIsCentroid and RearrangesDataset
