+++ src/mlpack/tests/kde_test.cpp
@@ -11,6 +11,7 @@
 
 #include <mlpack/methods/kde/kde.hpp>
 #include <mlpack/core/tree/binary_space_tree.hpp>
+#include <mlpack/core/tree/rectangle_tree.hpp>
 
 #include <boost/test/unit_test.hpp>
 #include "test_tools.hpp"
