### 2018-05-03 df4e030
Fix tree building
