### 2018-05-04 9fa129b
Fix uninitialized pointer
