### 2018-05-04 2cca9f5
Implement relative error tolerance
