### 2018-05-04 8398686
Implement Evaluate(Tree...)
Also add TreeAsArguments test
