### 2018-05-04 c1dedcc
Add methods to get and modify KDE parameters
