### 2018-05-04 1bd1eec
Add KDE copy constructor
