### 2018-05-04 ca36323
Add KDE operator=
