### 2018-05-05 #703 @Kevin-Mattheus-Moerman
@rcurtin minor point on the paper, can you please add DOI's to all references that have them. Thanks.
