### 2018-05-05 #703 @Kevin-Mattheus-Moerman
Review invitation. @MikeInnes @betatim @rasbt @cxhernandez @mlgill @rhiever would you be interested in reviewing this submission?

[-->project repo :floppy_disk:  <--](https://github.com/mlpack/mlpack)
[--> Paper  :page_facing_up: <--](https://github.com/openjournals/joss-papers/blob/joss.00703/joss.00703/10.21105.joss.00703.pdf)
