### 2018-05-05 7355d21
Added DET tests
