+++ src/mlpack/tests/main_tests/det_test.cpp
@@ -1,0 +1,381 @@
+/**
+  * @file det_test.cpp
+  * @author Manish Kumar
+  *
+  * Test mlpackMain() of det_main.cpp
+  *
+  * mlpack is free software; you may redistribute it and/or modify it under the
+  * terms of the 3-clause BSD license.  You should have received a copy of the
+  * 3-clause BSD license along with mlpack.  If not, see
+  * http://www.opensource.org/licenses/BSD-3-Clause for more information.
+  */
+#include <string>
+
+#define BINDING_TYPE BINDING_TYPE_TEST
+
+static const std::string testName = "DET";
+
+#include <mlpack/core.hpp>
+#include <mlpack/core/util/mlpack_main.hpp>
+#include "test_helper.hpp"
+#include <mlpack/methods/det/det_main.cpp>
+
+#include <boost/test/unit_test.hpp>
+#include "../test_tools.hpp"
+
+using namespace mlpack;
+
+struct DETTestFixture
+{
+ public:
+  DETTestFixture()
+  {
+    // Cache in the options for this program.
+    CLI::RestoreSettings(testName);
+  }
+
+  ~DETTestFixture()
+  {
+    // Clear the settings.
+    bindings::tests::CleanMemory();
+    CLI::ClearSettings();
+  }
+};
+
+BOOST_FIXTURE_TEST_SUITE(DETMainTest, DETTestFixture);
+
+/**
+ * Check that number of output training_set_estimates and number of input data
+ * points are equal.
+ */
+BOOST_AUTO_TEST_CASE(DETOutputDimensionTest)
+{
+  arma::mat trainingData;
+  if (!data::Load("iris.csv", trainingData))
+    BOOST_FAIL("Unable to load dataset iris.csv!");
+
+  arma::mat testData;
+  if (!data::Load("iris_test.csv", testData))
+    BOOST_FAIL("Unable to load dataset iris.csv!");
+
+  // Input data.
+  SetInputParam("training", trainingData);
+  SetInputParam("test", testData);
+
+  mlpackMain();
+
+  // Check the training_set_estimates has 100 points.
+  BOOST_REQUIRE_EQUAL(CLI::GetParam<arma::mat>("training_set_estimates").n_rows,
+                      1);
+  BOOST_REQUIRE_EQUAL(CLI::GetParam<arma::mat>("training_set_estimates").n_cols,
+                      trainingData.n_cols);
+
+  // Check the test_set_estimates has 40 points.
+  BOOST_REQUIRE_EQUAL(CLI::GetParam<arma::mat>("test_set_estimates").n_rows, 1);
+  BOOST_REQUIRE_EQUAL(CLI::GetParam<arma::mat>("test_set_estimates").n_cols,
+                      testData.n_cols);
+}
+
+/**
+ * Ensure that max_leaf_size & min_leaf_size are always positive and number of
+ * folds is always non-negative.
+ */
+BOOST_AUTO_TEST_CASE(DETParamBoundTest)
+{
+  arma::mat trainingData;
+  if (!data::Load("iris.csv", trainingData))
+    BOOST_FAIL("Unable to load dataset iris.csv!");
+
+  // Test for max_leaf_size.
+
+  SetInputParam("training", trainingData);
+  SetInputParam("max_leaf_size", (int) 0);
+
+  Log::Fatal.ignoreInput = true;
+  BOOST_REQUIRE_THROW(mlpackMain(), std::runtime_error);
+  Log::Fatal.ignoreInput = false;
+
+  bindings::tests::CleanMemory();
+
+  // Test for min_leaf_size.
+
+  SetInputParam("training", trainingData);
+  SetInputParam("min_leaf_size", (int) 0);
+
+  Log::Fatal.ignoreInput = true;
[…]
