### 2018-05-05 e9a652a
Add KDE move constructor
