### 2018-05-05 d6e1dbf
Convert Fedora instructions to python2, but note that we could also use python 3.
