+++ src/mlpack/methods/ann/layer/convolution.hpp
@@ -61,7 +61,7 @@ class Convolution
    * @param dH Stride of filter application in the y direction.
    * @param padW Padding width of the input.
    * @param padH Padding height of the input.
-   * @param inputWidth The widht of the input data.
+   * @param inputWidth The width of the input data.
    * @param inputHeight The height of the input data.
    */
   Convolution(const size_t inSize,
@@ -257,7 +257,7 @@ class Convolution
 
     for (size_t i = 0; i < input.n_slices; ++i)
     {
-      Pad<double>(input.slice(i), wPad, hPad, output.slice(i));
+      Pad<eT>(input.slice(i), wPad, hPad, output.slice(i));
     }
   }
 
