+++ src/mlpack/tests/ann_layer_test.cpp
@@ -1439,6 +1439,133 @@ BOOST_AUTO_TEST_CASE(GradientBatchNormLayerTest)
 }
 
 /**
+ * Simple Transposed Convolution layer test.
+ */
+BOOST_AUTO_TEST_CASE(SimpleTransposedConvolutionLayerTest)
+{
+  arma::mat output, input, delta;
+
+  TransposedConvolution<> module1(1, 1, 3, 3, 1, 1, 0, 0, 4, 4);
+  // Test the Forward function.
+  input = arma::linspace<arma::colvec>(0, 15, 16);
+  module1.Parameters() = arma::mat(9, 1, arma::fill::zeros);
+  module1.Parameters()(0) = 1.0;
+  module1.Parameters()(8) = 2.0;
+  module1.Reset();
+  module1.Forward(std::move(input), std::move(output));
+  // Value calculated using tensorflow.nn.conv2d_transpose()
+  BOOST_REQUIRE_EQUAL(arma::accu(output), 360.0);
+
+  // Test the Backward function.
+  module1.Backward(std::move(input), std::move(output), std::move(delta));
+  BOOST_REQUIRE_EQUAL(arma::accu(delta), 720);
+
+  TransposedConvolution<> module2(1, 1, 4, 4, 1, 1, 2, 2, 5, 5);
+  // Test the forward function.
+  input = arma::linspace<arma::colvec>(0, 24, 25);
+  module2.Parameters() = arma::mat(16, 1, arma::fill::zeros);
+  module2.Parameters()(0) = 1.0;
+  module2.Parameters()(3) = 1.0;
+  module2.Parameters()(6) = 1.0;
+  module2.Parameters()(9) = 1.0;
+  module2.Parameters()(12) = 1.0;
+  module2.Parameters()(15) = 2.0;
+  module2.Reset();
+  module2.Forward(std::move(input), std::move(output));
+  // Value calculated using tensorflow.nn.conv2d_transpose()
+  BOOST_REQUIRE_EQUAL(arma::accu(output), 2100.0);
+
+  // Test the backward function.
+  module2.Backward(std::move(input), std::move(output), std::move(delta));
+  BOOST_REQUIRE_EQUAL(arma::accu(delta), 7740);
+
+  TransposedConvolution<> module3(1, 1, 3, 3, 1, 1, 1, 1, 5, 5);
+  // Test the forward function.
+  input = arma::linspace<arma::colvec>(0, 24, 25);
+  module3.Parameters() = arma::mat(9, 1, arma::fill::zeros);
+  module3.Parameters()(1) = 2.0;
+  module3.Parameters()(2) = 4.0;
+  module3.Parameters()(3) = 3.0;
+  module3.Parameters()(8) = 1.0;
+  module3.Reset();
+  module3.Forward(std::move(input), std::move(output));
+  // Value calculated using tensorflow.nn.conv2d_transpose()
+  BOOST_REQUIRE_EQUAL(arma::accu(output), 3000.0);
+
+  // Test the backward function.
+  module3.Backward(std::move(input), std::move(output), std::move(delta));
+  BOOST_REQUIRE_EQUAL(arma::accu(delta), 21480);
+
+  TransposedConvolution<> module4(1, 1, 3, 3, 1, 1, 2, 2, 5, 5);
+  // Test the forward function.
+  input = arma::linspace<arma::colvec>(0, 24, 25);
+  module4.Parameters() = arma::mat(9, 1, arma::fill::zeros);
+  module4.Parameters()(2) = 2.0;
+  module4.Parameters()(4) = 4.0;
+  module4.Parameters()(6) = 6.0;
+  module4.Parameters()(8) = 8.0;
+  module4.Reset();
[…]

[…]
