### 2018-05-06 #703 @rasbt
i would be happy to do that. Timeline-wise, I think I will be able to get started on that later this week or next weekend.
