### 2018-05-07 #726 @whedon
[REVIEW]: mlpack 3: a fast, flexible machine learning library
**Submitting author:** @rcurtin (<a href="http://orcid.org/0000-0002-9903-8214">Ryan Curtin</a>)
**Repository:** <a href="https://github.com/mlpack/mlpack" target ="_blank">https://github.com/mlpack/mlpack</a>
**Version:** v3.0.2
