### 2018-05-07 #703 @Kevin-Mattheus-Moerman
Okay @rcurtin @mlgill @rasbt MikeInnes I've started the review issue over at #726
