### 2018-05-07 #726 @rcurtin
Sounds good---do let me know if there is anything required from me during the reviews.

In openjournals/joss#406, I opened an issue about this paper.  Authors 4 and 5 (Sumedh and Yannis) should have no affiliation, but I could not figure out a way to get this behavior.
