### 2018-05-07 #726 @rasbt
Sounds good!

@rcurtin I will try to go through the checklist chronologically throughout this and next week. If I have comments/suggestions, I will open those as issues in the mlpack GitHub repo as suggested by Joss.
