+++ src/mlpack/methods/lsh/lsh_main.cpp
@@ -203,7 +203,7 @@ static void mlpackMain()
         std::move(CLI::GetParam<arma::Mat<size_t>>("true_neighbors"));
 
     if (trueNeighbors.n_rows != neighbors.n_rows ||
-          trueNeighbors.n_cols != neighbors.n_cols)
+        trueNeighbors.n_cols != neighbors.n_cols)
         Log::Fatal << "The true neighbors file must have the same number of "
             << "values as the set of neighbors being queried!" << endl;
 
