### 2018-05-09 69f145d
Added Pendulum continuous environment of OpenAI Gym
