### 2018-05-10 731d8e8
Remove const requirement from KernelType
