### 2018-05-10 3052579
Use unsafe_col to speed up KDE score
