### 2018-05-10 5f503e2
Implement Layer Normalization
