### 2018-05-10 d172dbf
Handle kernel and metric as KDE member objects
