### 2018-05-10 12d3a7b
Add BSD license.
