+++ src/mlpack/tests/main_tests/approx_kfn_test.cpp
@@ -3,6 +3,11 @@
  * @author Namrata Mukhija
  *
  * Test mlpackMain() of approx_kfn_main.cpp.
+ *
+ * mlpack is free software; you may redistribute it and/or modify it under the
+ * terms of the 3-clause BSD license.  You should have received a copy of the
+ * 3-clause BSD license along with mlpack.  If not, see
+ * http://www.opensource.org/licenses/BSD-3-Clause for more information.
  */
 #include <string>
 
