+++ doc/guide/build.hpp
@@ -22,7 +22,7 @@ to build mlpack on Windows, see <a
 href="https://keon.io/mlpack-on-windows/">Keon's excellent tutorial</a>.
 
 You can download the latest mlpack release from here:
-<a href="http://www.mlpack.org/files/mlpack-3.0.0.tar.gz">mlpack-3.0.0</a>
+<a href="http://www.mlpack.org/files/mlpack-3.0.1.tar.gz">mlpack-3.0.1</a>
 
 @section build_simple Simple Linux build instructions
 
@@ -30,9 +30,9 @@ Assuming all dependencies are installed in the system, you can run the commands
 below directly to build and install mlpack.
 
 @code
-$ wget http://www.mlpack.org/files/mlpack-3.0.0.tar.gz
-$ tar -xvzpf mlpack-3.0.0.tar.gz
-$ mkdir mlpack-3.0.0/build && cd mlpack-3.0.0/build
+$ wget http://www.mlpack.org/files/mlpack-3.0.1.tar.gz
+$ tar -xvzpf mlpack-3.0.1.tar.gz
+$ mkdir mlpack-3.0.1/build && cd mlpack-3.0.1/build
 $ cmake ../
 $ make -j4  # The -j is the number of cores you want to use for a build.
 $ sudo make install
@@ -50,8 +50,8 @@ configure mlpack.
 First we should unpack the mlpack source and create a build directory.
 
 @code
-$ tar -xvzpf mlpack-3.0.0.tar.gz
-$ cd mlpack-3.0.0
+$ tar -xvzpf mlpack-3.0.1.tar.gz
+$ cd mlpack-3.0.1
 $ mkdir build
 @endcode
 
