+++ doc/guide/python_quickstart.hpp
@@ -31,9 +31,9 @@ build and install mlpack.  You can copy-paste the commands into your shell.
 @code{.sh}
 sudo apt-get install libboost-all-dev g++ cmake libarmadillo-dev python-pip wget
 sudo pip install cython setuptools distutils numpy pandas
-wget http://www.mlpack.org/files/mlpack-3.0.0.tar.gz
-tar -xvzpf mlpack-3.0.0.tar.gz
-mkdir -p mlpack-3.0.0/build/ && cd mlpack-3.0.0/build/
+wget http://www.mlpack.org/files/mlpack-3.0.1.tar.gz
+tar -xvzpf mlpack-3.0.1.tar.gz
+mkdir -p mlpack-3.0.1/build/ && cd mlpack-3.0.1/build/
 cmake ../ && make -j4 && sudo make install
 @endcode
 
