### 2018-05-11 380d437
Fix LaTeX commands.
