### 2018-05-11 8382083
Add KDE custom kernel and metric constructor
