### 2018-05-11 4824d06
Add KDE breadth-first support
