### 2018-05-11 f2dfcd1
Fixed style errors
