### 2018-05-11 a87261d
Add stochastic PCA policy.
