+++ src/mlpack/methods/pca/decomposition_policies/stochastic_method.hpp
@@ -1,0 +1,318 @@
+/**
+ * @file stochastic_method.hpp
+ * @author Marcus Edel
+ *
+ * Implementation of the stochastic method for use in the Principal Components
+ * Analysis method.
+ *
+ * mlpack is free software; you may redistribute it and/or modify it under the
+ * terms of the 3-clause BSD license.  You should have received a copy of the
+ * 3-clause BSD license along with mlpack.  If not, see
+ * http://www.opensource.org/licenses/BSD-3-Clause for more information.
+ */
+
+#ifndef MLPACK_METHODS_PCA_DECOMPOSITION_POLICIES_STOACHASTIC_METHOD_HPP
+#define MLPACK_METHODS_PCA_DECOMPOSITION_POLICIES_STOACHASTIC_METHOD_HPP
+
+#include <mlpack/prereqs.hpp>
+#include <mlpack/core/optimizers/sgd/sgd.hpp>
+#include <mlpack/core/optimizers/adam/adam.hpp>
+#include <mlpack/core/optimizers/svrg/svrg.hpp>
+
+#include <mlpack/core/util/sfinae_utility.hpp>
+
+namespace mlpack {
+namespace pca {
+
+//! Detect an StepSize() method.
+HAS_MEM_FUNC(StepSize, HasStepSize);
+
+/**
+ * Implementation of the Stochastic Policy. The stochastic approximation
+ * algorithm is an iterative algorithm, where in each iteration a single
+ * sampled point is used to perform an update as done in Stochastic Gradient
+ * Descent. For PCA this means iteratively using vectors sampled from a
+ * distribution to update the subspace.
+ *
+ * For more information, see the following.
+ *
+ * @code
+ * @inproceedings{Musco2015,
+ *   author    = {Arora, Raman and Cotter, Andrew and Livescu, Karen
+ *                and Srebro, Nathan},
+ *   title     = {Stochastic optimization for PCA and PLS.},
+ *   booktitle = {Allerton Conference},
+ *   pages     = {861-868},
+ *   year      = {2012},
+ * }
+ * @endcode
+ */
+template<typename OptimizerType>
+class StochasticPolicyType
+{
+ public:
+  /**
+   * Use the stochastic gradient method to perform the principal components
+   * analysis (PCA).
+   *
+   * @param optimizer Instantiated optimizer used to update the subspace.
+   */
+  StochasticPolicyType(const OptimizerType& optimizer = OptimizerType()) :
+      optimizer(optimizer)
+  {
+    /* Nothing to do here. */
+  }
+
+  /**
+   * Apply Principal Component Analysis to the provided data set using the
+   * exact SVD method.
+   *
+   * @param data Data matrix.
+   * @param centeredData Centered data matrix.
+   * @param transformedData Matrix to put results of PCA into.
+   * @param eigVal Vector to put eigenvalues into.
+   * @param eigvec Matrix to put eigenvectors (loadings) into.
+   * @param rank Rank of the decomposition.
+   */
+  void Apply(const arma::mat& data,
+             const arma::mat& centeredData,
+             arma::mat& transformedData,
+             arma::vec& eigVal,
+             arma::mat& eigvec,
+             const size_t rank)
+  {
+    PCAFunction f(centeredData);
+
+    // Adjust the step size if necessary.
+    SetStepSize(optimizer);
+
+    eigvec = arma::randn(centeredData.n_rows, rank);
+    optimizer.Optimize(f, eigvec);
+
+    // Project the samples to the principals.
+    transformedData = arma::trans(eigvec) * centeredData;
+
+    // Approximate eigenvalues and eigenvectors using Rayleigh–Ritz method.
+    arma::mat u, v;
+    arma::svd_econ(u, eigVal, v, transformedData);
+
+    // Now we must square the singular values to get the eigenvalues.
[…]
