### 2018-05-11 542347e
Add stochastic PCA policy test.
