+++ src/mlpack/tests/pca_test.cpp
@@ -14,6 +14,7 @@
 #include <mlpack/methods/pca/pca.hpp>
 #include <mlpack/methods/pca/decomposition_policies/exact_svd_method.hpp>
 #include <mlpack/methods/pca/decomposition_policies/quic_svd_method.hpp>
+#include <mlpack/methods/pca/decomposition_policies/stochastic_method.hpp>
 #include <mlpack/methods/pca/decomposition_policies/randomized_svd_method.hpp>
 #include <mlpack/methods/pca/decomposition_policies/randomized_block_krylov_method.hpp>
 
@@ -34,7 +35,8 @@ using namespace mlpack::distribution;
 template<typename DecompositionPolicy>
 void ArmaComparisonPCA(
     const bool scaleData = false,
-    const DecompositionPolicy& decomposition = DecompositionPolicy())
+    const DecompositionPolicy& decomposition = DecompositionPolicy(),
+    const double tolerance = 0.0001)
 {
   arma::mat coeff, coeff1, score, score1;
   arma::vec eigVal, eigVal1;
@@ -52,7 +54,7 @@ void ArmaComparisonPCA(
     if (eigVal[i] == 0.0)
       BOOST_REQUIRE_SMALL(eigVal1[i], 1e-15);
     else
-      BOOST_REQUIRE_CLOSE(eigVal[i], eigVal1[i], 0.0001);
+      BOOST_REQUIRE_CLOSE(eigVal[i], eigVal1[i], tolerance);
   }
 }
 
@@ -69,6 +71,8 @@ void PCADimensionalityReduction(
   mat data("1 0 2 3 9;"
            "5 2 8 4 8;"
            "6 7 3 1 8");
+
+  std::cout << "data: \n" << data << std::endl;
 
   // Now run PCA to reduce the dimensionality.
   size_t trial = 0;
@@ -89,6 +93,10 @@ void PCADimensionalityReduction(
   }
 
   BOOST_REQUIRE_EQUAL(success, true);
+
+  std::cout << "reduced data: \n" << data << std::endl;
+
+  // exit(0);
 
   // Compare with correct results.
   mat correct("-1.53781086 -3.51358020 -0.16139887 -1.87706634  7.08985628;"
@@ -205,6 +213,24 @@ BOOST_AUTO_TEST_CASE(ArmaComparisonRandomizedBlockKrylovPCATest)
 BOOST_AUTO_TEST_CASE(ArmaComparisonRandomizedPCATest)
 {
   ArmaComparisonPCA<RandomizedSVDPolicy>();
+}
+
+
+/**
+ * Compare the output of stochastic SGD PCA implementation with Armadillo's.
+ */
+BOOST_AUTO_TEST_CASE(ArmaComparisonStochasticSGDPCATest)
+{
+  ArmaComparisonPCA<StochasticSGDPolicy >();
+
+}
+
+/**
+ * Compare the output of stochastic Adam PCA implementation with Armadillo's.
+ */
+BOOST_AUTO_TEST_CASE(ArmaComparisonStochasticAdamPCATest)
+{
+  ArmaComparisonPCA<StochasticAdamPolicy >();
 }
 
 /**
