### 2018-05-11 e386c9d
Minor style fixes (comments, indentation, empty lines).
