### 2018-05-12 e2a6573
Fixed style and other errors
