### 2018-05-12 8bc9491
Added pendulum tests
