### 2018-05-14 ae2ef99
Switched to Camel Case and fixed some style issues
