### 2018-05-14 0d33805
Minor fix for type
