### 2018-05-14 a072209
merge-multiply layer added
