+++ src/mlpack/methods/ann/layer/layer.hpp
@@ -22,6 +22,7 @@
 #include "linear.hpp"
 #include "linear_no_bias.hpp"
 #include "lstm.hpp"
+#include "multiply_merge.hpp"
 #include "gru.hpp"
 #include "fast_lstm.hpp"
 #include "recurrent.hpp"
