+++ src/mlpack/methods/ann/layer/multiply_merge.hpp
@@ -1,0 +1,172 @@
+/**
+ * @file multiply_merge.hpp
+ * @author Haritha Nair
+ *
+ * Definition of the MultiplyMerge module which multiplies the output of the
+ * given modules element-wise.
+ *
+ * mlpack is free software; you may redistribute it and/or modify it under the
+ * terms of the 3-clause BSD license.  You should have received a copy of the
+ * 3-clause BSD license along with mlpack.  If not, see
+ * http://www.opensource.org/licenses/BSD-3-Clause for more information.
+ */
+#ifndef MLPACK_METHODS_ANN_LAYER_MULTIPLY_MERGE_HPP
+#define MLPACK_METHODS_ANN_LAYER_MULTIPLY_MERGE_HPP
+
+#include <mlpack/prereqs.hpp>
+
+#include "../visitor/delete_visitor.hpp"
+#include "../visitor/delta_visitor.hpp"
+#include "../visitor/output_parameter_visitor.hpp"
+
+#include "layer_types.hpp"
+
+namespace mlpack {
+namespace ann /** Artificial Neural Network. */ {
+
+/**
+ * Implementation of the MultiplyMerge module class. The MultiplyMerge class
+ * multiplies the output of various modules element-wise.
+ *
+ * @tparam InputDataType Type of the input data (arma::colvec, arma::mat,
+ *         arma::sp_mat or arma::cube).
+ * @tparam OutputDataType Type of the output data (arma::colvec, arma::mat,
+ *         arma::sp_mat or arma::cube).
+ * @tparam CustomLayers Additional custom layers that can be added.
+ */
+template<
+    typename InputDataType = arma::mat,
+    typename OutputDataType = arma::mat,
+    typename... CustomLayers
+>
+class MultiplyMerge
+{
+ public:
+  /**
+   * Create the MultiplyMerge object using the specified parameters.
+   *
+   * @param model Expose the all network modules.
+   */
+  MultiplyMerge(const bool model = false);
+
+  //! Destructor to release allocated memory.
+  ~MultiplyMerge();
+
+  /**
+   * Ordinary feed forward pass of a neural network, evaluating the function
+   * f(x) by propagating the activity forward through f.
+   *
+   * @param input Input data used for evaluating the specified function.
+   * @param output Resulting output activation.
+   */
+  template<typename InputType, typename OutputType>
+  void Forward(const InputType&& /* input */, OutputType&& output);
+
+  /**
+   * Ordinary feed backward pass of a neural network, calculating the function
+   * f(x) by propagating x backwards trough f, using the results from the feed
+   * forward pass.
+   *
+   * @param input The propagated input activation.
+   * @param gy The backpropagated error.
+   * @param g The calculated gradient.
+   */
+  template<typename eT>
+  void Backward(const arma::Mat<eT>&& /* input */,
+                arma::Mat<eT>&& gy,
+                arma::Mat<eT>&& g);
+
+  /*
+   * Add a new module to the model.
+   *
+   * @param layer The Layer to be added to the model.
+   */
+  void Add(LayerTypes<CustomLayers...> layer) { network.push_back(layer); }
+
+  /*
+   * Add a new module to the model.
+   *
+   * @param layer The Layer to be added to the model.
+   */
+  template<typename LayerType>
+  void Add(const LayerType& layer) { network.push_back(new LayerType(layer)); }
+
+  /*
+   * Add a new module to the model.
+   *
+   * @param args The layer parameter.
+   */
+  template <class LayerType, class... Args>
+  void Add(Args... args) { network.push_back(new LayerType(args...)); }
+
+  //! Get the input parameter.
+  InputDataType const& InputParameter() const { return inputParameter; }
[…]
