### 2018-05-14 b625cba
Adjust parameter size and use alias inside conv operation.
