### 2018-05-15 1336322
Fixed the scope of power function and typo in tests
