### 2018-05-15 315839f
Added name to contributors
