### 2018-05-15 2d5c8e9
Fix Continuous Mountain Car environment
