### 2018-05-15 bcdcb47
addition to layer types
