### 2018-05-15 29371f6
Resolve failing Tests
