### 2018-05-16 ea69b9a
Fixed style issue
