+++ src/mlpack/methods/reinforcement_learning/environment/continuous_mountain_car.hpp
@@ -92,8 +92,10 @@ class ContinuousMountainCar
    *
    * @param positionMin Minimum legal position.
    * @param positionMax Maximum legal position.
+   * @param positionGoal Final target position.
    * @param velocityMin Minimum legal velocity.
    * @param velocityMax Maximum legal velocity.
+   * @param power Power generated by car.
    */
   ContinuousMountainCar(const double positionMin = -1.2,
                         const double positionMax = 0.6,
