### 2018-05-16 a30f5f8
Add explicit 'None' affiliations.
