### 2018-05-16 #726 @rcurtin
I've redone the paper PDF to have an explicit "None" affiliation for unaffiliated authors, but if openjournals/whedon#23 is merged I will update the paper.
