### 2018-05-16 937973b
add cf no_normalization
