### 2018-05-16 6cad42e
normalizationType constructor
