### 2018-05-17 0630ed3
add normalization cmakelist
