+++ src/mlpack/methods/cf/cf_impl.hpp
@@ -53,13 +53,16 @@ void ApplyFactorizer(FactorizerType& factorizer,
 /**
  * Construct the CF object using an instantiated factorizer.
  */
+template<typename NormalizationType>
 template<typename FactorizerType>
-CF::CF(const arma::mat& data,
+CF<NormalizationType>::CF(const arma::mat& data,
        FactorizerType factorizer,
        const size_t numUsersForSimilarity,
-       const size_t rank) :
+       const size_t rank,
+       const NormalizationType normalization) :
     numUsersForSimilarity(numUsersForSimilarity),
-    rank(rank)
+    rank(rank),
+    normalization(normalization)
 {
   // Validate neighbourhood size.
   if (numUsersForSimilarity < 1)
@@ -70,21 +73,26 @@ CF::CF(const arma::mat& data,
     this->numUsersForSimilarity = 5;
   }
 
+  // Normalize data and train.
+  normalization.Normalize(data);
   Train(data, factorizer);
 }
 
 /**
  * Construct the CF object using an instantiated factorizer.
  */
+template<typename NormalizationType>
 template<typename FactorizerType>
-CF::CF(const arma::sp_mat& data,
+CF<NormalizationType>::CF(const arma::sp_mat& data,
        FactorizerType factorizer,
        const size_t numUsersForSimilarity,
        const size_t rank,
+       const NormalizationType normalization,
        const typename std::enable_if_t<
            !FactorizerTraits<FactorizerType>::UsesCoordinateList>*) :
     numUsersForSimilarity(numUsersForSimilarity),
-    rank(rank)
+    rank(rank),
+    normalization(normalization)
 {
   // Validate neighbourhood size.
   if (numUsersForSimilarity < 1)
@@ -95,11 +103,14 @@ CF::CF(const arma::sp_mat& data,
     this->numUsersForSimilarity = 5;
   }
 
+  // Normalize data and train.
+  normalization.Normalize(data);
   Train(data, factorizer);
 }
 
+template<typename NormalizationType>
 template<typename FactorizerType>
-void CF::Train(const arma::mat& data, FactorizerType factorizer)
+void CF<NormalizationType>::Train(const arma::mat& data, FactorizerType factorizer)
 {
   CleanData(data, cleanedData);
 
@@ -125,8 +136,9 @@ void CF::Train(const arma::mat& data, FactorizerType factorizer)
   Timer::Stop("cf_factorization");
 }
 
+template<typename NormalizationType>
 template<typename FactorizerType>
-void CF::Train(const arma::sp_mat& data,
+void CF<NormalizationType>::Train(const arma::sp_mat& data,
                FactorizerType factorizer,
                const typename std::enable_if_t<!FactorizerTraits<
                    FactorizerType>::UsesCoordinateList>*)
@@ -154,8 +166,9 @@ void CF::Train(const arma::sp_mat& data,
 }
 
 //! Serialize the model.
+template<typename NormalizationType>
 template<typename Archive>
-void CF::serialize(Archive& ar, const unsigned int /* version */)
+void CF<NormalizationType>::serialize(Archive& ar, const unsigned int /* version */)
 {
   // This model is simple; just serialize all the members.  No special handling
   // required.
@@ -164,7 +177,272 @@ void CF::serialize(Archive& ar, const unsigned int /* version */)
   ar & BOOST_SERIALIZATION_NVP(w);
   ar & BOOST_SERIALIZATION_NVP(h);
   ar & BOOST_SERIALIZATION_NVP(cleanedData);
+  ar & BOOST_SERIALIZATION_NVP(normalization);
 }
+
+
+
+
+// Default CF constructor.
+template<typename NormalizationType>
[…]
