### 2018-05-17 e0f9abb
change CF to CF<>
