### 2018-05-17 40d399d
add normalization to cmakelist
