+++ src/mlpack/methods/cf/decomposition_policies/nmf_method.hpp
@@ -52,7 +52,7 @@ class NMFPolicy
              const double minResidue,
              const bool mit)
   {
-    if(mit)
+    if (mit)
     {
       amf::MaxIterationTermination iter(maxIterations);
 
