### 2018-05-19 5a94434
add overall mean normalization
