### 2018-05-19 2cffbb3
add user/item normalization
