+++ src/mlpack/methods/cf/normalization/item_mean_normalization.hpp
@@ -1,0 +1,132 @@
+/**
+ * @file item_mean_normalization.hpp
+ * @author Wenhao Huang
+ *
+ * This class performs item mean normalization on raw ratings. In another
+ * word, this class is used to remove global effect of item mean.
+ *
+ * mlpack is free software; you may redistribute it and/or modify it under the
+ * terms of the 3-clause BSD license.  You should have received a copy of the
+ * 3-clause BSD license along with mlpack.  If not, see
+ * http://www.opensource.org/licenses/BSD-3-Clause for more information.
+ */
+#ifndef MLPACK_METHODS_CF_NORMALIZATION_ITEM_MEAN_NORMALIZATION_HPP
+#define MLPACK_METHODS_CF_NORMALIZATION_ITEM_MEAN_NORMALIZATION_HPP
+
+#include <mlpack/prereqs.hpp>
+
+namespace mlpack {
+namespace cf {
+
+/**
+ * This normalization class performs item mean normalization on raw ratings.
+ */
+class ItemMeanNormalization
+{
+ public:
+  // Empty constructor.
+  ItemMeanNormalization() { }
+
+  /**
+   * Normalize the data by subtracting item mean from each of existing rating.
+   *
+   * @param data Input dataset in the form of coordinate list.
+   */
+  void Normalize(arma::mat& data)
+  {
+    const size_t itemNum = arma::max(data.row(1)) + 1;
+    // Should we use overall mean if an item has no rating?
+    itemMean = arma::vec(itemNum, arma::fill:zeros);
+    // Number of ratings for each item.
+    arma::vec ratingNum(itemNum, arma::fill::zeros);
+
+    // Sum ratings for each item.
+    data.each_col([&](vec& datapoint) {
+      const size_t item = (size_t) datapoint(1);
+      const double rating = datapoint(2);
+      itemMean(item) += rating;
+      ratingNum(item) += 1;
+    });
+
+    // Calculate item mean and subtract item mean from ratings.
+    // Set item mean to 0 if the item has no rating.
+    for (int i = 0; i < itemNum; i++)
+      if ((size_t) itemNum != 0)
+        itemMean(i) /= itemNum(i);
+    data.each_col([&](vec& datapoint) {
+      const size_t item = (size_t) datapoint(1);
+      datapoint(2) -= itemMean(item);
+    });
+  }
+
+  /**
+   * Normalize the data by subtracting item mean from each of existing rating.
+   *
+   * @param cleanedData Sparse matrix data.
+   */
+  void Normalize(arma::sp_mat& cleanedData)
+  {
+    itemMean = arma::mean(cleanedData, 1);
+
+    arma::sp_mat::iterator it = cleanedData.begin();
+    arma::sp_mat::iterator it_end = cleanedData.end(); 
+    for (; it != it_end; it++)
+      *it = *it - itemMean(it.row());
+  }
+
+  /**
+   * Denormalize computed rating by adding item mean.
+   *
+   * @param user User ID.
+   * @param item Item ID.
+   * @param rating Computed rating before denormalization.
+   */
+  double Denormalize(const int /* user */,
+                     const int item,
+                     const double rating) const
+  {
+    return rating + itemMean(item);
+  }
+
+  /**
+   * Denormalize computed rating by adding item mean.
+   *
+   * @param combinations User/Item combinations.
+   * @param predictions Predicted ratings for each user/item combination.
+   */
+  void Denormalize(const arma::Mat<size_t>& combinations,
+                   arma::vec& predictions) const
+  {
+    for (int i = 0; i < predictions.n_elem; i++)
+    {
[…]
