### 2018-05-20 457dcaf
add z-score normalization
