### 2018-05-20 2a0f2cd
add combined normalization
