### 2018-05-20 mlpack/mlpack#1403 @mlgill
[JOSS] DOIs needed in paper
@rcurtin I am reviewing the mlpack submission for the JOSS submission. It looks like DOIs are missing from the paper.
