### 2018-05-20 mlpack/mlpack#1404 @mlgill
[JOSS] Add dependencies and installation instructions to main website
@rcurtin I see the list of dependencies on GitHub [here](https://github.com/mlpack/mlpack/blob/master/README.md#3-dependencies) followed by installation instructions. Would it be possible to add this information (or a link to the GitHub dependencies) and on the main website [here](http://www.mlpack.org/) somewhere after the download links, etc?
