### 2018-05-20 #726 @mlgill
Have requested dependency and installation information be added to main website, issue is [here](https://github.com/mlpack/mlpack/issues/1404).
