### 2018-05-20 mlpack/mlpack#1405 @rasbt
[JOSS] LICENSE.txt and COPYRIGHT.txt slightly out of sync
Hi, there,

going through the JOSS checklist for the software review, I saw that the LICENSE.txt and COPYRIGHT.txt are slightly out of sync:

LICENSE.txt:
> Copyright (c) 2007-2016, mlpack contributors (see COPYRIGHT.txt)

COPYRIGHT.txt:
> Copyright 2008-2018, Ryan Curtin <ryan@ratml.org>
> ...

For the version that will get archived with the JOSS paper, it's maybe best to update the License file with the most up to date copyright info
