### 2018-05-20 #726 @mlgill
@arfon @rcurtin I noticed the version on GitHub has been bumped to 3.0.1 (our checklist says 3.0.0). Is there somewhere else that I should verify that the update has been made?
