### 2018-05-20 mlpack/mlpack#1404 @rasbt
I found a list of the dependencies on the documentation website following the links:

http://www.mlpack.org => docs => read the mlpack tutorials => Building mlpack From Source => Dependencies of mlpack

However, I agree that it would be nice to have a more obvious place for installation instructions and requirements. Ideally as a link in the navbar on http://www.mlpack.org or via a link like "Installation Instructions" on the http://www.mlpack.org/docs page.
