### 2018-05-20 #726 @Kevin-Mattheus-Moerman
@mlgill thanks for pointing that out. We'll request to get that fixed before acceptance. By the way please address editorial comments in this issue to me directly.
