### 2018-05-20 mlpack/mlpack#1406 @rasbt
[JOSS] Recommended apt-get installation installs old version (mlpack 2.0.1)?
I followed the instructions recommended in the docs via apt-get, and installed it as follows (on Ubuntu 16.04 LTS)

```
sudo apt-get upgrade
sudo apt-get update
sudo apt-get install libmlpack-dev
```

When I then look at the files installed, via `dpkg -L libmlpack-dev`, I see that these list

>  * This file is part of mlpack 2.0.1.

I then checked via `apt list --installed | grep libmlpack-dev`

and saw that it seems to be the old version indeed (`libmlpack-dev/xenial,now 2.0.1-1 amd64 [installed]`)

I am wondering if this is an issue with compatible versions for my system, or is the 3.0.1 version not available via apt-get, yet? If so, can the 3.0.1 version for this review currently only be installed from source?
