### 2018-05-20 #726 @rasbt
@rcurtin have begun the review and bumped into an issue with the installation instructions (https://github.com/mlpack/mlpack/issues/1406), would be nice if you could take a look at it some time. Other than that, I found a very minor issue with the copyright info in the copyright & license files being out of sync (https://github.com/mlpack/mlpack/issues/1405).
