### 2018-05-20 mlpack/mlpack#1406 @rasbt
Related, the command-line program reference documentation refers to version 2.2.5 (http://www.mlpack.org/docs/mlpack-3.0.1/man.html):

> Below is a list of the command-line executables mlpack 2.2.5 provides, with links to the documentation for each executable. This documentation may also be accessed with the --help parameter or through the man pages provided with your distribution of mlpack.

Is this just a documentation bug or is the mlpack 3.0.0 functionality currently only available for the C++ library and language (e.g. Python) bindings?
