### 2018-05-21 mlpack/mlpack#1406 @rcurtin
Thanks for pointing out the issue---I fixed http://www.mlpack.org/docs/mlpack-3.0.1/man.html; it was an oversight during the release process.

For apt, we don't control the mlpack package in the repositories.  I can see that 3.0.1 has been pushed to the upstream Debian unstable repositories: https://packages.debian.org/sid/libmlpack-dev, but it hasn't made it to Ubuntu yet.  Some time ago I submitted a request for Ubuntu to update the mlpack package on older Ubuntu versions but I received no response.
