### 2018-05-21 83b697d
Update DOIs where possible.
