+++ src/mlpack/methods/lmnn/lmnn_impl.hpp
@@ -37,7 +37,7 @@ LMNN<MetricType, OptimizerType>::LMNN(const arma::mat& dataset,
 template<typename MetricType, typename OptimizerType>
 void LMNN<MetricType, OptimizerType>::LearnDistance(arma::mat& outputMatrix)
 {
- // See if we were passed an initialized matrix.
+  // See if we were passed an initialized matrix.
   if ((outputMatrix.n_rows != dataset.n_rows) ||
       (outputMatrix.n_cols != dataset.n_rows))
     outputMatrix.eye(dataset.n_rows, dataset.n_rows);
