### 2018-05-21 mlpack/mlpack#1405 @rcurtin
Fixed in #1408---thanks for pointing it out.  Just a minor oversight.
