### 2018-05-21 mlpack/mlpack#1403 @rcurtin
I added the DOIs that I could find for the references in #1408, but not every reference had a DOI.  I am not the most familiar person in the world with DOIs, so please double-check my work (83b697d) and let me know if I didn't do what you were hoping for.  Thanks!
