+++ src/mlpack/methods/lmnn/lmnn_function_impl.hpp
@@ -55,12 +55,11 @@ void LMNNFunction<MetricType>::Shuffle()
 
   dataset = std::move(newDataset);
   labels = std::move(newLabels);
-}
 
-//! Calculate outer product of vector.
-arma::mat OuterProduct(arma::vec& diff)
-{
-  return diff * arma::trans(diff);
+  // Re-calculate target neighbors as indices changed.
+  // TODO: Devise a better way of doing this.
+  Constraints constraint(dataset, labels, k);
+  constraint.TargetNeighbors(targetNeighbors);
 }
 
 //! Evaluate cost over whole dataset.
@@ -150,18 +149,22 @@ double LMNNFunction<MetricType>::Evaluate(const arma::mat& coordinates,
 
 //! Compute gradient over whole dataset.
 template<typename MetricType>
+template<typename GradType>
 void LMNNFunction<MetricType>::Gradient(const arma::mat& coordinates,
-                                        arma::mat& gradient)
+                                        GradType& gradient)
 {
   gradient.zeros(dataset.n_rows, dataset.n_rows);
+
+  arma::mat cij = arma::zeros(dataset.n_rows, dataset.n_rows);
+  arma::mat cil = arma::zeros(dataset.n_rows, dataset.n_rows);
 
   for (size_t i = 0; i < dataset.n_rows; i++)
   {
     for (size_t j = 0; j < k ; j++)
     {
       // Calculate gradient due to target neighbors.
-      arma::vec cij = dataset.col(i) - dataset.col(targetNeighbors(j, i));
-      gradient += (1 - regularization) * OuterProduct(cij);
+      arma::vec diff = dataset.col(i) - dataset.col(targetNeighbors(j, i));
+      cij += diff * arma::trans(diff);
     }
 
     for (size_t j = 0; j < k ; j++)
@@ -169,18 +172,21 @@ void LMNNFunction<MetricType>::Gradient(const arma::mat& coordinates,
       for (size_t l = 0; l < k ; l++)
       {
         // Calculate gradient due to triplets.
-        arma::vec cij = dataset.col(i) - dataset.col(targetNeighbors(j, i));
-        arma::vec cil = dataset.col(i) - dataset.col(impostors(l, i));
+        arma::vec diff = dataset.col(i) - dataset.col(targetNeighbors(j, i));
+        cil += diff * arma::trans(diff);
 
-        gradient += regularization * (OuterProduct(cij) - OuterProduct(cil));
+        diff = dataset.col(i) - dataset.col(impostors(l, i));
+        cil -= diff * arma::trans(diff);
       }
     }
   }
+
+  gradient = (1 - regularization) * cij + regularization * cil;
 }
 
 //! Compute gradient over a batch of data points.
 template<typename MetricType>
-template <typename GradType>
+template<typename GradType>
 void LMNNFunction<MetricType>::Gradient(const arma::mat& coordinates,
                                         const size_t begin,
                                         GradType& gradient,
@@ -188,13 +194,16 @@ void LMNNFunction<MetricType>::Gradient(const arma::mat& coordinates,
 {
   gradient.zeros(dataset.n_rows, dataset.n_rows);
 
+  arma::mat cij = arma::zeros(dataset.n_rows, dataset.n_rows);
+  arma::mat cil = arma::zeros(dataset.n_rows, dataset.n_rows);
+
   for (size_t i = begin; i < begin + batchSize; i++)
   {
     for (size_t j = 0; j < k ; j++)
     {
       // Calculate gradient due to target neighbors.
-      arma::vec cij = dataset.col(i) - dataset.col(targetNeighbors(j, i));
-      gradient += (1 - regularization) * OuterProduct(cij);
[…]
