+++ src/mlpack/methods/lmnn/lmnn_function.hpp
@@ -72,10 +72,12 @@ class LMNNFunction
    * matrix.  This is the non-separable implementation, where the objective
    * function is not decomposed into the sum of several objective functions.
    *
+   * @tparam GradType The type of the gradient out-param.
    * @param covariance Covariance matrix of Mahalanobis distance.
    * @param gradient Matrix to store the calculated gradient in.
    */
-  void Gradient(const arma::mat& covariance, arma::mat& gradient);
+  template<typename GradType>
+  void Gradient(const arma::mat& covariance, GradType& gradient);
 
   /**
    * Evaluate the gradient of the LMNN function for the given covariance
@@ -92,11 +94,46 @@ class LMNNFunction
    * @param batchSize Number of points to use for objective function.
    * @param gradient Matrix to store the calculated gradient in.
    */
-  template <typename GradType>
+  template<typename GradType>
   void Gradient(const arma::mat& covariance,
                 const size_t begin,
                 GradType& gradient,
                 const size_t batchSize = 1);
+
+  /**
+   * Evaluate the LMNN objective function together with gradient for the given
+   * covariance matrix.  This is the non-separable implementation, where the
+   * objective function is not decomposed into the sum of several objective
+   * functions.
+   *
+   * @tparam GradType The type of the gradient out-param.
+   * @param covariance Covariance matrix of Mahalanobis distance.
+   * @param gradient Matrix to store the calculated gradient in.
+   */
+  template<typename GradType>
+  void EvaluateWithGradient(const arma::mat& coordinates,
+                            GradType& gradient);
+
+  /**
+   * Evaluate the LMNN objective function together with gradient for the given
+   * covariance matrix on the given batch size, from a given initial point of
+   * the dataset. This is the separable implementation, where the objective
+   * function is decomposed into the sum of many objective functions, and
+   * here, only one of those constituent objective functions is returned.
+   * The type of the gradient parameter is a template
+   * argument to allow the computation of a sparse gradient.
+   *
+   * @tparam GradType The type of the gradient out-param.
+   * @param covariance Covariance matrix of Mahalanobis distance.
+   * @param begin Index of the initial point to use for objective function.
+   * @param batchSize Number of points to use for objective function.
+   * @param gradient Matrix to store the calculated gradient in.
+   */
+  template<typename GradType>
+  void EvaluateWithGradient(const arma::mat& covariance,
+                            const size_t begin,
+                            GradType& gradient,
+                            const size_t batchSize = 1);
 
   //! Return the initial point for the optimization.
   const arma::mat& GetInitialPoint() const { return initialPoint; }
