### 2018-05-23 757fe1c
Added test for NCA binding
