+++ src/mlpack/tests/main_tests/nca_test.cpp
@@ -1,0 +1,445 @@
+/**
+ * @file nca_test.cpp
+ * @author Yasmine Dumouchel
+ *
+ * Test mlpackMain() of nca_main.cpp.
+ *
+ * mlpack is free software; you may redistribute it and/or modify it under the
+ * terms of the 3-clause BSD license.  You should have received a copy of the
+ * 3-clause BSD license along with mlpack.  If not, see
+ * http://www.opensource.org/licenses/BSD-3-Clause for more information.
+ */
+
+#include <string>
+
+#define BINDING_TYPE BINDING_TYPE_TEST
+static const std::string testName = "nca";
+
+#include <mlpack/core.hpp>
+#include <mlpack/core/util/mlpack_main.hpp>
+#include <mlpack/core/metrics/lmetric.hpp>
+#include "test_helper.hpp"
+#include <mlpack/methods/nca/nca_main.cpp>
+#include <mlpack/core/optimizers/lbfgs/lbfgs.hpp>
+
+#include <boost/test/unit_test.hpp>
+#include "../test_tools.hpp"
+
+using namespace mlpack;
+
+struct NCATestFixture
+{
+ public:
+  NCATestFixture()
+  {
+    // Cache in the options for this program.
+    CLI::RestoreSettings(testName);
+  }
+
+  ~NCATestFixture()
+  {
+    // Clear the settings.
+    bindings::tests::CleanMemory();
+    CLI::ClearSettings();
+  }
+};
+
+BOOST_FIXTURE_TEST_SUITE(NCAMainTest, NCATestFixture);
+
+/**
+ * Ensure that, when labels are implicitily given with input,
+ * the last column is treated as labels and that we get the
+ * desired shape of output.
+ */
+BOOST_AUTO_TEST_CASE(NCAExplicitImplicitLabelsTest)
+{
+  // Random dataset.
+  arma::mat x;
+  x.randu(3, 6);
+
+  SetInputParam("input", std::move(x));
+
+  mlpackMain();
+
+  // Check that last row was treated as label by checking that
+  // the output has 1 less row.
+  BOOST_REQUIRE_EQUAL(CLI::GetParam<arma::mat>("output").n_rows, 2);
+  BOOST_REQUIRE_EQUAL(CLI::GetParam<arma::mat>("output").n_cols, 2);
+
+  // Reset Settings.
+  CLI::ClearSettings();
+  CLI::RestoreSettings(testName);
+
+  // Now check that when labels are explicitely given, the last column
+  // of input is not treated as labels.
+  arma::mat y              = "-0.1 -0.1 -0.1  0.1  0.1  0.1;"
+                             " 1.0  0.0 -1.0  1.0  0.0 -1.0 ";
+  arma::Row<size_t> labels = " 0    0    0    1    1    1   ";
+
+  SetInputParam("input", std::move(y));
+  SetInputParam("labels", std::move(labels));
+
+  mlpackMain();
+
+  // Check that final output has expected number of rows and colums.
+  BOOST_REQUIRE_EQUAL(CLI::GetParam<arma::mat>("output").n_rows, 2);
+  BOOST_REQUIRE_EQUAL(CLI::GetParam<arma::mat>("output").n_cols, 2);
+}
+
+/**
+ * Ensure that when we pass optimizer of type lbfgs, we also get the desired
+ * shape of output.
+ */
+BOOST_AUTO_TEST_CASE(NCALBFGSTest)
+{
+  arma::mat x;
+  x.randu(3, 100);
+  arma::Row<size_t> labels;
+  labels.zeros(3);
+
+  // Input random data points.
+  SetInputParam("input", std::move(x));
+  SetInputParam("labels", std::move(labels));
[…]
