### 2018-05-23 4851b6d
Added some tests - inspired by NCA
