+++ src/mlpack/tests/lmnn_test.cpp
@@ -1,0 +1,352 @@
+/**
+ * @file lmnn_test.cpp
+ * @author Manish Kumar
+ * Unit tests for Large Margin Nearest Neighbors and related code (including
+ * the constraints class).
+ *
+ * mlpack is free software; you may redistribute it and/or modify it under the
+ * terms of the 3-clause BSD license.  You should have received a copy of the
+ * 3-clause BSD license along with mlpack.  If not, see
+ * http://www.opensource.org/licenses/BSD-3-Clause for more information.
+ */
+#include <mlpack/core.hpp>
+#include <mlpack/core/metrics/lmetric.hpp>
+#include <mlpack/methods/lmnn/lmnn.hpp>
+#include <mlpack/core/optimizers/lbfgs/lbfgs.hpp>
+
+#include <boost/test/unit_test.hpp>
+#include "test_tools.hpp"
+
+using namespace mlpack;
+using namespace mlpack::metric;
+using namespace mlpack::lmnn;
+using namespace mlpack::optimization;
+
+
+BOOST_AUTO_TEST_SUITE(LMNNTest);
+
+//
+// Tests for the Constraints.
+//
+
+/**
+ * The target neighbors function should be correct.
+ * point.
+ */
+BOOST_AUTO_TEST_CASE(LMNNTargetNeighborsTest)
+{
+  // Useful but simple dataset with six points and two classes.
+  arma::mat dataset        = "-0.1 -0.1 -0.1  0.1  0.1  0.1;"
+                             " 1.0  0.0 -1.0  1.0  0.0 -1.0 ";
+  arma::Row<size_t> labels = " 0    0    0    1    1    1   ";
+
+  Constraints constraint(dataset, labels, 1);
+
+  //! Store target neighbors of data points.
+  arma::Mat<size_t> targetNeighbors;
+
+  constraint.TargetNeighbors(targetNeighbors);
+
+  BOOST_REQUIRE_EQUAL(targetNeighbors(0, 0), 1);
+  BOOST_REQUIRE_EQUAL(targetNeighbors(0, 1), 0);
+  BOOST_REQUIRE_EQUAL(targetNeighbors(0, 2), 1);
+  BOOST_REQUIRE_EQUAL(targetNeighbors(0, 3), 4);
+  BOOST_REQUIRE_EQUAL(targetNeighbors(0, 4), 3);
+  BOOST_REQUIRE_EQUAL(targetNeighbors(0, 5), 4);
+}
+
+/**
+ * The impostors function should be correct.
+ */
+BOOST_AUTO_TEST_CASE(LMNNImpostorsTest)
+{
+  // Useful but simple dataset with six points and two classes.
+  arma::mat dataset        = "-0.1 -0.1 -0.1  0.1  0.1  0.1;"
+                             " 1.0  0.0 -1.0  1.0  0.0 -1.0 ";
+  arma::Row<size_t> labels = " 0    0    0    1    1    1   ";
+
+  Constraints constraint(dataset, labels, 1);
+
+  //! Store impostors of data points.
+  arma::Mat<size_t> impostors;
+
+  constraint.Impostors(impostors);
+
[…]
