+++ src/mlpack/methods/lmnn/lmnn_function_impl.hpp
@@ -36,9 +36,10 @@ LMNNFunction<MetricType>::LMNNFunction(const arma::mat& dataset,
   // Initialize the initial learning point.
   initialPoint.eye(dataset.n_rows, dataset.n_rows);
 
-  // Initialize target neighbors.
+  // Initialize target neighbors & impostors.
   Constraints constraint(dataset, labels, k);
   constraint.TargetNeighbors(targetNeighbors);
+  constraint.Impostors(impostors);
 }
 
 //! Shuffle the dataset.
@@ -62,10 +63,30 @@ void LMNNFunction<MetricType>::Shuffle()
   constraint.TargetNeighbors(targetNeighbors);
 }
 
+inline void Projection(arma::mat& iterate)
+{
+  // Projection.
+  arma::vec eigval;
+  arma::mat eigvec;
+
+  arma::eig_sym(eigval, eigvec, iterate);
+
+  arma::uvec ind = arma::find(eigval > 0);
+  arma::mat diagEigVal = arma::diagmat(eigval);
+  iterate = eigvec.cols(ind) * diagEigVal.submat(ind, ind) *
+      arma::trans(eigvec.cols(ind));
+}
+
 //! Evaluate cost over whole dataset.
 template<typename MetricType>
 double LMNNFunction<MetricType>::Evaluate(const arma::mat& coordinates)
 {
+  // Hack
+  arma::mat& iterate = const_cast<arma::mat&>(coordinates);
+  // Apply projection.
+  if (arma::accu(iterate == initialPoint) != initialPoint.n_cols)
+    Projection(iterate);
+
   double cost = 0;
 
   // Apply metric over dataset.
@@ -110,6 +131,12 @@ double LMNNFunction<MetricType>::Evaluate(const arma::mat& coordinates,
                                           const size_t begin,
                                           const size_t batchSize)
 {
+  // Hack
+  arma::mat& iterate = const_cast<arma::mat&>(coordinates);
+  // Apply projection.
+  if (arma::accu(iterate == initialPoint) != initialPoint.n_cols)
+    Projection(iterate);
+
   double cost = 0;
 
   // Apply metric over dataset.
@@ -150,7 +177,7 @@ double LMNNFunction<MetricType>::Evaluate(const arma::mat& coordinates,
 //! Compute gradient over whole dataset.
 template<typename MetricType>
 template<typename GradType>
-void LMNNFunction<MetricType>::Gradient(const arma::mat& coordinates,
+void LMNNFunction<MetricType>::Gradient(const arma::mat& /* coordinates */,
                                         GradType& gradient)
 {
   gradient.zeros(dataset.n_rows, dataset.n_rows);
@@ -158,7 +185,7 @@ void LMNNFunction<MetricType>::Gradient(const arma::mat& coordinates,
   arma::mat cij = arma::zeros(dataset.n_rows, dataset.n_rows);
   arma::mat cil = arma::zeros(dataset.n_rows, dataset.n_rows);
 
-  for (size_t i = 0; i < dataset.n_rows; i++)
+  for (size_t i = 0; i < dataset.n_cols; i++)
   {
     for (size_t j = 0; j < k ; j++)
     {
@@ -187,7 +214,7 @@ void LMNNFunction<MetricType>::Gradient(const arma::mat& coordinates,
 //! Compute gradient over a batch of data points.
 template<typename MetricType>
 template<typename GradType>
-void LMNNFunction<MetricType>::Gradient(const arma::mat& coordinates,
+void LMNNFunction<MetricType>::Gradient(const arma::mat& /* coordinates */,
                                         const size_t begin,
                                         GradType& gradient,
                                         const size_t batchSize)
[…]
