### 2018-05-23 mlpack/mlpack#1404 @rcurtin
I added a link to "build instructions" on the http://www.mlpack.org/download.html page; since you already have to go to that page to download the source, I think it's reasonable to put it right below the links for downloading the source.

Let me know if you think that's reasonable---I'll go ahead close this issue since I think it is handled.

Thanks for pointing it out! :)
