### 2018-05-23 #726 @rcurtin
Hi there @mlgill and @rasbt, thanks for your comments.  I've addressed each of the issues you opened (mlpack/mlpack#1403, mlpack/mlpack#1404, mlpack/mlpack#1405, mlpack/mlpack#1406)---let me know if everything is handled to your satisfaction.

You are right that we have released 3.0.1 (it's a pretty incremental release).  We can update the version in the paper, but maybe it is good to wait until the review process is done---we may release 3.0.2 (another bugfix release) between now and then.  Personally given the lack of serious changes between 3.0.1 and 3.0.0, I think it would be nicer to release the paper as 3.0.0, since that's a nicer "round" number.  But I don't know the JOSS policy on such things, so let me know what you'd like me to do or if anything is needed from my end. :)
