### 2018-05-23 #726 @Kevin-Mattheus-Moerman
@rcurtin our policy is that the version number should reflect the state of the reviewed software. So if you implement fixes in 3.0.2 relevant to the review then that should be the version number for the paper.
