### 2018-05-23 #726 @rcurtin
Your call, either way is fine with me. :)
