### 2018-05-23 mlpack/mlpack#1406 @rasbt
I see. I'd say it would be good to add a small note about that until it gets updated. Since Ubuntu is a pretty widely adopted distro these days, it could safe some users from some misunderstanding
