### 2018-05-23 f5bee6e
Add explicit test for label size.
The previous test worked in debugging mode but not in release mode because it
depended on Armadillo exceptions, which aren't thrown unless in debugging mode.
