### 2018-05-23 e18d61d
Apply linear_scan to the rest of the tests that use SGD.
Also modify the normalization test a bit:

 * Apply normalization only in the second run.  This is because of a subtle
   issue that probably should eventually be fixed---if we explicitly set a flag
   to false, the CLI object still will return true for HasParam() (but not
   GetParam<bool>()), and many of the bindings use HasParam() preferentially
   instead of GetParam() for flags.

 * Apply a small tolerance to reduce the amount of time that the test takes
   significantly.
