### 2018-05-24 770c807
resolve minor issues
