### 2018-05-25 a419476
add average interpolation
