### 2018-05-26 9213cff
change distances to similarities
