### 2018-05-26 7af0ab7
add similarity based interpolation
