+++ src/mlpack/methods/cf/interpolation_policies/similarity_interpolation.hpp
@@ -38,8 +38,8 @@ class SimilarityInterpolation
   {
     if (similarities.n_elem == 0)
     {
-      Log::Fatal << "Require: similarities.n_elem > 0. There should be at least "
-          << "one neighbor!" << std::endl;
+      Log::Fatal << "Require: similarities.n_elem > 0. There should be at "
+          << "least one neighbor!" << std::endl;
     }
 
     double similaritiesSum = arma::sum(similarites);
