### 2018-05-27 b9006c0
debugged matrix error
