### 2018-05-27 3fb772e
debug emptyctortest
