### 2018-05-28 b294b31
Additional Improvements and Remove Debugging Comments
