+++ src/mlpack/methods/ann/ffn_impl.hpp
@@ -57,7 +57,6 @@ template<typename OutputLayerType, typename InitializationRuleType,
 void FFN<OutputLayerType, InitializationRuleType, CustomLayers...>::ResetData(
     arma::mat predictors, arma::mat responses)
 {
-  std::cout << "Enter FFN ResetData:" << std::endl;
   numFunctions = responses.n_cols;
   this->predictors = std::move(predictors);
   this->responses = std::move(responses);
@@ -66,7 +65,6 @@ void FFN<OutputLayerType, InitializationRuleType, CustomLayers...>::ResetData(
 
   if (!reset)
     ResetParameters();
-  std::cout << "Exit FFN ResetData:" << std::endl;
 }
 
 template<typename OutputLayerType, typename InitializationRuleType,
@@ -77,7 +75,6 @@ void FFN<OutputLayerType, InitializationRuleType, CustomLayers...>::Train(
       arma::mat responses,
       OptimizerType& optimizer)
 {
-  std::cout << "Enter FFN Train(1):" << std::endl;
   ResetData(std::move(predictors), std::move(responses));
 
   // Train the model.
@@ -87,7 +84,6 @@ void FFN<OutputLayerType, InitializationRuleType, CustomLayers...>::Train(
 
   Log::Info << "FFN::FFN(): final objective of trained model is " << out
       << "." << std::endl;
-  std::cout << "Exit FFN Train:" << std::endl;
 }
 
 template<typename OutputLayerType, typename InitializationRuleType,
@@ -96,7 +92,6 @@ template<typename OptimizerType>
 void FFN<OutputLayerType, InitializationRuleType, CustomLayers...>::Train(
     arma::mat predictors, arma::mat responses)
 {
-  std::cout << "Enter FFN Train(2):" << std::endl;
   ResetData(std::move(predictors), std::move(responses));
 
   OptimizerType optimizer;
@@ -108,7 +103,6 @@ void FFN<OutputLayerType, InitializationRuleType, CustomLayers...>::Train(
 
   Log::Info << "FFN::FFN(): final objective of trained model is " << out
       << "." << std::endl;
-  std::cout << "Exit FFN Train:" << std::endl;
 }
 
 template<typename OutputLayerType, typename InitializationRuleType,
@@ -116,7 +110,6 @@ template<typename OutputLayerType, typename InitializationRuleType,
 void FFN<OutputLayerType, InitializationRuleType, CustomLayers...>::Forward(
     arma::mat inputs, arma::mat& results)
 {
-  std::cout << "Enter FFN Forward(1):" << std::endl;
   if (parameter.is_empty())
     ResetParameters();
 
@@ -129,7 +122,6 @@ void FFN<OutputLayerType, InitializationRuleType, CustomLayers...>::Forward(
   currentInput = std::move(inputs);
   Forward(std::move(currentInput));
   results = boost::apply_visitor(outputParameterVisitor, network.back());
-  std::cout << "Exit FFN Forward:" << std::endl;
 }
 
 template<typename OutputLayerType, typename InitializationRuleType,
@@ -137,7 +129,6 @@ template<typename OutputLayerType, typename InitializationRuleType,
 double FFN<OutputLayerType, InitializationRuleType, CustomLayers...>::Backward(
     arma::mat targets, arma::mat& gradients)
 {
-  std::cout << "Enter FFN Backward(1):" << std::endl;
   double res = outputLayer.Forward(std::move(boost::apply_visitor(
       outputParameterVisitor, network.back())), std::move(targets));
 
@@ -149,7 +140,7 @@ double FFN<OutputLayerType, InitializationRuleType, CustomLayers...>::Backward(
   Backward();
   ResetGradients(gradients);
[…]

[…]
