### 2018-05-29 473d187
regSVD debugging
