### 2018-05-30 24c8af0
add lmetric_search
