+++ src/mlpack/methods/cf/neighbor_search_polices/lmetric_search.hpp
@@ -1,0 +1,70 @@
+/**
+ * @file lmetric_search.hpp
+ * @author Wenhao Huang
+ *
+ * Search with L_p distance.
+ *
+ * mlpack is free software; you may redistribute it and/or modify it under the
+ * terms of the 3-clause BSD license.  You should have received a copy of the
+ * 3-clause BSD license along with mlpack.  If not, see
+ * http://www.opensource.org/licenses/BSD-3-Clause for more information.
+ */
+#ifndef MLPACK_METHODS_CF_LMETRIC_SEARCH_HPP
+#define MLPACK_METHODS_CF_LMETRIC_SEARCH_HPP
+
+#include <mlpack/prereqs.hpp>
+#include <mlpack/methods/neighbor_search/neighbor_search.hpp>
+#include <mlpack/core/metrics/lmetric.hpp>
+
+namespace mlpack {
+namespace cf {
+
+/**
+ * Search with L_p distance.
+ *
+ * @tparam TPower Power of metric.
+ */
+template<int TPower>
+class LMetricSearch
+{
+ public:
+  using NeighborSearchType = neighbor::NeighborSearch<
+      neighbor::NearestNeighborSort,
+      metric::LMetric<TPower, true>>;
+
+  /**
+   * @param Set of reference points.
+   */
+  LMetricSearch(const arma::mat& referenceSet) : neighborSearch(referenceSet)
+  { }
+
+  /**
+   * Given a set of query points, find the nearest k neighbors, and return
+   * similarites. Similarities are less than or equal to one.
+   *
+   * @param query A set of query points.
+   * @param k Number of neighbors to search.
+   * @param neighbors Nearest neighbors.
+   * @param similarites Similarities between query point and its neighbors.
+   */
+  void Search(const arma::mat& query, const size_t k,
+      arma::Mat<size_t>& neighbors, arma::mat& similarities)
+  {
+    neighbor.Search(query, k, neighbors, similarities);
+    
+    // Calculate similarities from L_p distance. We restrict that similarities
+    // are not larger than one.
+    similarites = 1.0 / (1.0 + similarities);
+  }
+
+ private:
+  //! NeighborSearch object.
+  NeighborSearchType neighborSearch;
+};
+
+using EuclideanSearch = LMetricSearch<2>;
+
+} // namespace cf
+} // namespace mlpack
+
+#endif
