+++ src/mlpack/methods/cf/neighbor_search_polices/CMakeLists.txt
@@ -1,15 +1,8 @@
 # Define the files we need to compile
 # Anything not in this list will not be compiled into mlpack.
 set(SOURCES
-  cf.hpp
-  cf_impl.hpp
-  cf.cpp
-  svd_wrapper.hpp
-  svd_wrapper_impl.hpp
+  lmetric_search.hpp
 )
-
-add_subdirectory(interpolation_policies)
-add_subdirectory(neighbor_search_policies)
 
 # Add directory name to sources.
 set(DIR_SRCS)
@@ -19,6 +12,3 @@ endforeach()
 # Append sources (with directory name) to list of all mlpack sources (used at
 # the parent scope).
 set(MLPACK_SRCS ${MLPACK_SRCS} ${DIR_SRCS} PARENT_SCOPE)
-
-add_cli_executable(cf)
-add_python_binding(cf)
