### 2018-05-30 70680c7
Rectify Gradient
