### 2018-05-31 #726 @Kevin-Mattheus-Moerman
@mlgill @rasbt thanks for your comments so far. I understand @rcurtin has been addressing some of your comments. Can you give an update as to where we stand? Thanks :rocket:
