### 2018-05-31 ac54f6f
resolve failing tests
