### 2018-05-31 #726 @MikeInnes
I've checked off what I can verify, and things look generally good, but I haven't tried to run things locally yet. Are there build instructions оr binaries for OS X, or is this a Linux-only project?
