### 2018-05-31 #726 @rcurtin
Builds fine on OS X.  An old version of mlpack (2.2.5) is available via Homebrew, but it's not too hard to build---should be just `cmake ../ && make && make install` basically; see http://mlpack.org/docs/mlpack-3.0.1/doxygen/build.html.  The Linux build instructions should also work fine for OS X.
