### 2018-05-31 c37df2f
Handle training sets that don't divide evenly by k.
