### 2018-05-31 cd37c05
Add ShuffleData() with weights.
