### 2018-05-31 b71251d
Add 'shuffle' parameter to KFoldCV.
