+++ src/mlpack/core/cv/k_fold_cv_impl.hpp
@@ -26,57 +26,9 @@ KFoldCV<MLAlgorithm,
         PredictionsType,
         WeightsType>::KFoldCV(const size_t k,
                               const MatType& xs,
-                              const PredictionsType& ys) :
-    KFoldCV(Base(), k, xs, ys)
-{ /* Nothing left to do. */ }
-
-template<typename MLAlgorithm,
-         typename Metric,
-         typename MatType,
-         typename PredictionsType,
-         typename WeightsType>
-KFoldCV<MLAlgorithm,
-        Metric,
-        MatType,
-        PredictionsType,
-        WeightsType>::KFoldCV(const size_t k,
-                              const MatType& xs,
                               const PredictionsType& ys,
-                              const size_t numClasses) :
-    KFoldCV(Base(numClasses), k, xs, ys)
-{ /* Nothing left to do. */ }
-
-template<typename MLAlgorithm,
-         typename Metric,
-         typename MatType,
-         typename PredictionsType,
-         typename WeightsType>
-KFoldCV<MLAlgorithm,
-        Metric,
-        MatType,
-        PredictionsType,
-        WeightsType>::KFoldCV(const size_t k,
-                              const MatType& xs,
-                              const data::DatasetInfo& datasetInfo,
-                              const PredictionsType& ys,
-                              const size_t numClasses) :
-    KFoldCV(Base(datasetInfo, numClasses), k, xs, ys)
-{ /* Nothing left to do. */ }
-
-template<typename MLAlgorithm,
-         typename Metric,
-         typename MatType,
-         typename PredictionsType,
-         typename WeightsType>
-KFoldCV<MLAlgorithm,
-        Metric,
-        MatType,
-        PredictionsType,
-        WeightsType>::KFoldCV(const size_t k,
-                              const MatType& xs,
-                              const PredictionsType& ys,
-                              const WeightsType& weights) :
-    KFoldCV(Base(), k, xs, ys, weights)
+                              const bool shuffle) :
+    KFoldCV(Base(), k, xs, ys, shuffle)
 { /* Nothing left to do. */ }
 
 template<typename MLAlgorithm,
@@ -92,8 +44,8 @@ KFoldCV<MLAlgorithm,
                               const MatType& xs,
                               const PredictionsType& ys,
                               const size_t numClasses,
-                              const WeightsType& weights) :
-    KFoldCV(Base(numClasses), k, xs, ys, weights)
+                              const bool shuffle) :
+    KFoldCV(Base(numClasses), k, xs, ys, shuffle)
 { /* Nothing left to do. */ }
 
 template<typename MLAlgorithm,
@@ -110,8 +62,62 @@ KFoldCV<MLAlgorithm,
                               const data::DatasetInfo& datasetInfo,
                               const PredictionsType& ys,
                               const size_t numClasses,
-                              const WeightsType& weights) :
-    KFoldCV(Base(datasetInfo, numClasses), k, xs, ys, weights)
+                              const bool shuffle) :
+    KFoldCV(Base(datasetInfo, numClasses), k, xs, ys, shuffle)
+{ /* Nothing left to do. */ }
+
+template<typename MLAlgorithm,
+         typename Metric,
+         typename MatType,
+         typename PredictionsType,
+         typename WeightsType>
+KFoldCV<MLAlgorithm,
+        Metric,
+        MatType,
+        PredictionsType,
+        WeightsType>::KFoldCV(const size_t k,
+                              const MatType& xs,
+                              const PredictionsType& ys,
+                              const WeightsType& weights,
[…]
