+++ src/mlpack/core/cv/k_fold_cv.hpp
@@ -69,10 +69,14 @@ class KFoldCV
    * @param xs Data points to cross-validate on.
    * @param ys Predictions (labels for classification algorithms and responses
    *     for regression algorithms) for each data point.
+   * @param shuffle Whether or not to shuffle the data and predictions before
+   *     performing cross-validation.  Shuffling will be performed before every
+   *     call to Evaluate().
    */
   KFoldCV(const size_t k,
           const MatType& xs,
-          const PredictionsType& ys);
+          const PredictionsType& ys,
+          const bool shuffle = true);
 
   /**
    * This constructor can be used for multiclass classification algorithms.
@@ -81,11 +85,15 @@ class KFoldCV
    * @param xs Data points to cross-validate on.
    * @param ys Labels for each data point.
    * @param numClasses Number of classes in the dataset.
+   * @param shuffle Whether or not to shuffle the data and predictions before
+   *     performing cross-validation.  Shuffling will be performed before every
+   *     call to Evaluate().
    */
   KFoldCV(const size_t k,
           const MatType& xs,
           const PredictionsType& ys,
-          const size_t numClasses);
+          const size_t numClasses,
+          const bool shuffle = true);
 
   /**
    * This constructor can be used for multiclass classification algorithms that
@@ -96,12 +104,16 @@ class KFoldCV
    * @param datasetInfo Type information for each dimension of the dataset.
    * @param ys Labels for each data point.
    * @param numClasses Number of classes in the dataset.
+   * @param shuffle Whether or not to shuffle the data and predictions before
+   *     performing cross-validation.  Shuffling will be performed before every
+   *     call to Evaluate().
    */
   KFoldCV(const size_t k,
           const MatType& xs,
           const data::DatasetInfo& datasetInfo,
           const PredictionsType& ys,
-          const size_t numClasses);
+          const size_t numClasses,
+          const bool shuffle = true);
 
   /**
    * This constructor can be used for regression and binary classification
@@ -112,11 +124,15 @@ class KFoldCV
    * @param ys Predictions (labels for classification algorithms and responses
    *     for regression algorithms) for each data point.
    * @param weights Observation weights (for boosting).
+   * @param shuffle Whether or not to shuffle the data and predictions before
+   *     performing cross-validation.  Shuffling will be performed before every
+   *     call to Evaluate().
    */
   KFoldCV(const size_t k,
           const MatType& xs,
           const PredictionsType& ys,
-          const WeightsType& weights);
+          const WeightsType& weights,
+          const bool shuffle = true);
 
   /**
    * This constructor can be used for multiclass classification algorithms that
@@ -127,12 +143,16 @@ class KFoldCV
    * @param ys Labels for each data point.
    * @param numClasses Number of classes in the dataset.
    * @param weights Observation weights (for boosting).
+   * @param shuffle Whether or not to shuffle the data and predictions before
+   *     performing cross-validation.  Shuffling will be performed before every
+   *     call to Evaluate().
    */
   KFoldCV(const size_t k,
           const MatType& xs,
           const PredictionsType& ys,
           const size_t numClasses,
-          const WeightsType& weights);
+          const WeightsType& weights,
+          const bool shuffle = true);
 
   /**
    * This constructor can be used for multiclass classification algorithms that
@@ -144,13 +164,17 @@ class KFoldCV
    * @param ys Labels for each data point.
    * @param numClasses Number of classes in the dataset.
    * @param weights Observation weights (for boosting).
[…]
