### 2018-05-31 3a62cf2
Safer handling of sparse matrix values arrays.
This fixes #1411.
