### 2018-06-01 #726 @rasbt
@Kevin-Mattheus-Moerman

> Can you give an update as to where we stand?

I am still in the process of running the software on examples (testing the functionality) but I should be able to wrap this review up on/shortly after the weekend
