### 2018-06-01 6a301d7
test time reduction
