### 2018-06-01 85e1c3b
added sampling layer
