+++ src/mlpack/methods/ann/layer/CMakeLists.txt
@@ -75,6 +75,8 @@ set(SOURCES
   recurrent_attention_impl.hpp
   reinforce_normal.hpp
   reinforce_normal_impl.hpp
+  sampling.hpp
+  sampling_impl.hpp
   select.hpp
   select_impl.hpp
   sequential.hpp
