### 2018-06-01 fe0db81
add support for all types
