### 2018-06-01 mlpack/mlpack#1416 @rasbt
[JOSS] how to set up the mlpack Python bindings
I followed the installation instructions and everything went super smoothly thus far! :)

I can use the binaries from within the bin folder, and also the tests all pass:

```bash
~/mlpack-3.0.1/build$ bin/mlpack_test
Running 1351 test cases...

...
warning: solve(): system appears singular (rcond: 1.31214e-18); attempting approximate solution

warning: solve(): system appears singular (rcond: 1.31214e-18); attempting approximate solution

*** No errors detected
```

I am wondering now what step would be required to enable the Python bindings? Would I just need to add `~/mlpack-3.0.1/src/mlpack/bindings/python/mlpack` to my Python `PATH`? It would be nice if that could be briefly explained in the documentation (maybe in the "building from source" docs).
