### 2018-06-01 mlpack/mlpack#1416 @rcurtin
Honestly I'd say `make install` is the easiest way to do it, like the build directions suggest.  If you really want to do it manually in the build directory it's hard enough that I wanted to avoid adding it to the documentation... from the build directory:

```
export LD_LIBRARY_PATH=lib/
export PYTHONPATH=src/mlpack/bindings/python/
```

If you `make install` it will put it in `/usr/local/lib/pythonX.Y/site-packages/` which should automatically be in the `PYTHONPATH`.

As I write this though, I do think you are right, it would be useful to add a section to the build documentation mentioning how to work with the Python bindings without installing.
