### 2018-06-01 c726c32
Remove trainingSubsetSize and clarify comments.
