### 2018-06-01 63ef265
Rename variables for clarity.
