### 2018-06-01 cc8b31a
Expose Shuffle() to users.
