### 2018-06-01 mlpack/mlpack#1416 @rasbt
> If you make install it will put it in /usr/local/lib/pythonX.Y/site-packages/ which should automatically be in the PYTHONPATH.

Hm, I used that, but for some reason, it didn't add anything to the site-packages directory (checked both the python 2.7 and 3.5 dirs), which is why I assumed a manual step might be necessary (I used `sudo make install`)

```python
>>> import mlpack
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
ImportError: No module named mlpack
```

> As I write this though, I do think you are right, it would be useful to add a section to the build documentation mentioning how to work with the Python bindings without installing. I'll do that when I have a chance.

Thanks, that would be nice.
