### 2018-06-01 mlpack/mlpack#1416 @rcurtin
That's strange. Mind posting the output of `make install` (it will be long...)?  The Python bindings are pretty fiddly and I've found that different systems behave pretty differently.

(The Python bindings are also available via pip for manylinux platforms but where's the fun in that?? :))
