### 2018-06-01 mlpack/mlpack#1416 @rcurtin
Hmm it seems like the Python bindings aren't actually referenced in that install message.  Does 'make python' work?  It seems like CMake may not have configured the project to build the Python bindings, which could happen if the right Python dependencies aren't available.
