### 2018-06-01 mlpack/mlpack#1416 @rasbt
> Does 'make python' work?

no, it doesn't seem to work either

```
make: *** No rule to make target 'python'.  Stop.
```

For the whole installation procedure, I used a fresh Ubuntu install (16.04 LTS) and followed the docs step by step, including

```
$ sudo apt-get install libboost-math-dev libboost-program-options-dev
  libboost-test-dev libboost-serialization-dev libarmadillo-dev binutils-dev
  python-pandas python-numpy cython python-setuptools
```

Usually, I use the Miniconda Python distribution, however, for this review, I thought it would be better to follow the installation step by step without any dependencies installed that are not listed.

Do you think the Cmake step doesn't work with the system's Python, and is there any good way to check that?
