### 2018-06-01 mlpack/mlpack#1416 @rcurtin
Actually, I bet the Cython version is too old:
https://packages.ubuntu.com/xenial/cython

The Cython version needs to be 0.24 or newer, so you could try on Ubuntu 17.10, but unfortunately 16.04 is pretty old at this point (or nobody has backported a newer version yet).
