### 2018-06-01 mlpack/mlpack#1416 @rasbt
You are right!

> CMake Warning at CMake/FindPythonModule.cmake:44 (message):
  Unsuitable version of Python module Cython (0.24 or greater required).

Will try a newer Ubuntu version then
