### 2018-06-01 d278658
Implement DCGAN Test
