+++ src/mlpack/methods/ann/layer/convolution_impl.hpp
@@ -208,10 +208,9 @@ void Convolution<
 
       if (padW != 0 || padH != 0)
       {
-        gTemp.slice(inMap + batchCount * inSize) += output.submat(
-            rotatedFilter.n_rows / 2, rotatedFilter.n_cols / 2,
-            rotatedFilter.n_rows / 2 + gTemp.n_rows - 1,
-            rotatedFilter.n_cols / 2 + gTemp.n_cols - 1);
+        gTemp.slice(inMap + batchCount * inSize) += output.submat(padW, padH,
+            padW + gTemp.n_rows - 1,
+            padH + gTemp.n_cols - 1);
       }
       else
       {
@@ -288,10 +287,9 @@ void Convolution<
           (gradientTemp.n_rows < output.n_rows &&
           gradientTemp.n_cols < output.n_cols))
       {
-        gradientTemp.slice(outMapIdx) += output.submat(output.n_rows / 2,
-            output.n_cols / 2,
-            output.n_rows / 2 + gradientTemp.n_rows - 1,
-            output.n_cols / 2 + gradientTemp.n_cols - 1);
+        gradientTemp.slice(outMapIdx) += output.submat(padW, padH,
+            padW + gradientTemp.n_rows - 1,
+            padH + gradientTemp.n_cols - 1);
       }
       else
       {
