### 2018-06-02 mlpack/mlpack#1416 @rasbt
Just got one step further ... using Ubuntu 18 seemed to have solved the Cython dependency issue. Sorry, but I know face the next issue. It seems to be a library linking issue when I try to import mlpack in Python:

```python
>>> import mlpack
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/usr/local/lib/python2.7/dist-packages/mlpack-3.0.1-py2.7-linux-x86_64.egg/mlpack/__init__.py", line 11, in <module>
    from .test_python_binding import test_python_binding
ImportError: libmlpack.so.3: cannot open shared object file: No such file or directory
```

I just `grep`ed through the `cmake`, `make`, and `make install` outputs and couldn't find anything suspicious there; I pasted the relevant lines below. Have you encountered this or a similar issue in the past and have any tips/recommendation for resolving this?

```
sudo cmake ../
-- cotire 1.7.10 loaded.
-- Armadillo libraries: /usr/lib/libarmadillo.so
-- Boost version: 1.65.1
-- Found the following Boost libraries:
--   program_options
--   unit_test_framework
--   serialization
-- Regenerating arma_config.hpp.
-- Found Python: /usr/bin/python
-- Could NOT find Doxygen (missing: DOXYGEN_EXECUTABLE)
CMake Warning at CMakeLists.txt:487 (message):
  txt2man not found; man pages will not be generated.

-- Configuring done
-- Generating done
-- Build files have been written to: /home/ubuntu/mlpack-3.0.1/build
ubuntu@ip-172-31-14-44:~/mlpack-3.0.1/build$ sudo make
[  1%] Updating arma_config.hpp (if necessary)
-- Regenerating arma_config.hpp.
[  1%] Built target mlpack_arma_config
Moving header files to include/mlpack/
[  1%] Built target mlpack_headers
[  2%] Generating mlpack.pc (pkg-config) file.
[  2%] Built target pkgconfig
[  2%] Building CXX precompiled header src/mlpack/cotire/mlpack_CXX_prefix.hxx.gch

...

[ 16%] Linking CXX shared library ../../lib/libmlpack.so
[ 16%] Built target mlpack
running build
running build_py
copying /home/ubuntu/mlpack-3.0.1/build/src/mlpack/bindings/python/mlpack/__init__.py -> build/lib.linux-x86_64-2.7/mlpack
copying /home/ubuntu/mlpack-3.0.1/build/src/mlpack/bindings/python/mlpack/matrix_utils.py -> build/lib.linux-x86_64-2.7/mlpack
[ 16%] Built target python_copy
Scanning dependencies of target generate_pyx_sparse_coding
[ 16%] Building CXX object src/mlpack/methods/sparse_coding/CMakeFiles/

...
```

```
~/mlpack-3.0.1/build$ sudo make
...

[ 15%] Building CXX object src/mlpack/CMakeFiles/mlpack.dir/methods/sparse_autoencoder/sparse_autoencoder_function.cpp.o
[ 15%] Building CXX object src/mlpack/CMakeFiles/mlpack.dir/methods/sparse_autoencoder/maximal_inputs.cpp.o
[ 15%] Building CXX object src/mlpack/CMakeFiles/mlpack.dir/methods/sparse_coding/sparse_coding.cpp.o
[ 16%] Linking CXX shared library ../../lib/libmlpack.so
[ 16%] Built target mlpack
running build
running build_py
copying /home/ubuntu/mlpack-3.0.1/build/src/mlpack/bindings/python/mlpack/__init__.py -> build/lib.linux-x86_64-2.7/mlpack
copying /home/ubuntu/mlpack-3.0.1/build/src/mlpack/bindings/python/mlpack/matrix_utils.py -> build/lib.linux-x86_64-2.7/mlpack
[ 16%] Built target python_copy
...
```

```
~/mlpack-3.0.1/build$ sudo make install
[  1%] Updating arma_config.hpp (if necessary)
[  1%] Built target mlpack_arma_config

...

-- Installing: /usr/local/include/mlpack/methods/softmax_regression/softmax_regression.hpp
-- Installing: /usr/local/include/mlpack/methods/softmax_regression/softmax_regression_impl.hpp
-- Installing: /usr/local/include/mlpack/prereqs.hpp
-- Installing: /usr/local/include/mlpack/core.hpp
-- Installing: /usr/local/lib/libmlpack.so.3.0
-- Up-to-date: /usr/local/lib/libmlpack.so.3
-- Up-to-date: /usr/local/lib/libmlpack.so
running install
running bdist_egg
running egg_info
...
```

PS: `apt-get install pkg-config` could be added to the dependency list, i.e., in the lines

```
# apt-get install libboost-math-dev libboost-program-options-dev
  libboost-test-dev libboost-serialization-dev libarmadillo-dev binutils-dev
  python-pandas python-numpy cython python-setuptools
```
since it doesn't come with Ubuntu by default.
