### 2018-06-02 mlpack/mlpack#1416 @rcurtin
If it's installed to `/usr/local/`, on some systems (including Ubuntu) this is not in the default linker search path.  So you'll need to `export LD_LIBRARY_PATH=/usr/local/lib/` since that's where libmlpack.so.3 is.

pkg-config isn't a requirement for building, though, it's optional---so if it's not there, no problem.  (You just can't use pkg-config later to link against mlpack, for instance.)
