### 2018-06-02 mlpack/mlpack#1416 @rcurtin
(more details in the README at the end of section 4: https://github.com/mlpack/mlpack#4-building-mlpack-from-source)
