### 2018-06-02 mlpack/mlpack#1416 @rasbt
Good point, thanks! I usually use CentOS, and I think it's the default there neither (prob. why I have that line looming around in my .bashrc).

Anyways it seems to work now :). Could you add the last few lines of Section 4 to the online docs as well? I think that's very helpful info for users since Ubuntu is one of the recommended (or "example") distros. Don't want to be too picky, but then I can check off the "Installation: Does installation proceed as outlined in the documentation?" review item in good consciousness :)
