### 2018-06-02 mlpack/mlpack#1418 @rasbt
[JOSS] Documentation typo in Adaboost code example
In the Python Bindings docs of Adaboost (http://www.mlpack.org/docs/mlpack-3.0.1/python/adaboost.html), the following code example is provided:

    >>> adaboost(training=data, weak_learner='perceptron')
    >>> model = output['output']

I believe this should be

    >>> adaboost(training=data, weak_learner='perceptron')
    >>> model = output['output_model']

based on the description at the very bottom: (`output_model (mlpack.AdaBoostModelType): Output trained AdaBoost model.`)?
