### 2018-06-02 mlpack/mlpack#1416 @rcurtin
That's a fair change and I would agree with it.  I'll incorporate it with the other changes I said I would do earlier.
