### 2018-06-02 mlpack/mlpack#1418 @rcurtin
You're right about the lack of `output =`, this is a bug in the auto-generated documentation.  It always helps to have an extra set of eyes---thank you for pointing this out.  The `output` -> `output_model` is a specific error in the AdaBoost code.
