### 2018-06-02 mlpack/mlpack#1418 @rasbt
>  It always helps to have an extra set of eyes

I agree :). Coincidentally, not too long ago, I submitted a ML toolkit library to JOSS as well (mlxtend) and was bumping into similar issues with the autogenerated docs that the reviewer found
