### 2018-06-02 #726 @rasbt
To provide a more concrete heads-up (and a reference for me to stay on track): The installation eventually went smoothly and the software works as expected.

I was able to check off most of the things. The remaining items are

- Version number: I assume we should wait until the other reviewers finished the technical review in case there will be additional changes requested
-  Installation: just two tiny changes that are currently being addressed https://github.com/mlpack/mlpack/issues/1406 and https://github.com/mlpack/mlpack/issues/1416
- Functionality documentation: also here, only two minor documentation fixes remaining, https://github.com/mlpack/mlpack/issues/1419 and https://github.com/mlpack/mlpack/issues/1418
