### 2018-06-02 8ae9745
Optimized some parts
