+++ src/mlpack/methods/lmnn/constraints_impl.hpp
@@ -76,6 +76,47 @@ void Constraints::TargetNeighbors(arma::Mat<size_t>& outputMatrix)
   }
 }
 
+// Calculates k similar labeled nearest neighbors  on a
+// batch of data points.
+void Constraints::TargetNeighbors(arma::Mat<size_t>& outputMatrix,
+                                  const size_t begin,
+                                  const size_t batchSize)
+{
+  arma::mat subDataset = dataset.cols(begin, begin + batchSize - 1);
+  arma::Row<size_t> sublabels = labels.cols(begin, begin + batchSize - 1);
+
+  arma::Row<size_t> uniqueLabels = arma::unique(labels);
+
+  // KNN instance.
+  neighbor::KNN knn;
+
+  arma::Mat<size_t> neighbors;
+  arma::mat distances;
+
+  // Vectors to store indices.
+  arma::uvec indexSame;
+  arma::uvec subIndexSame;
+
+  for ( size_t i = 0; i < uniqueLabels.n_cols; i++)
+  {
+    // Calculate Target Neighbors.
+    indexSame = arma::find(labels == uniqueLabels[i]);
+    subIndexSame = arma::find(sublabels == uniqueLabels[i]);
+
+    // Perform KNN search with same class points as both reference
+    // set and query set.
+    knn.Train(dataset.cols(indexSame));
+    knn.Search(subDataset.cols(subIndexSame), k, neighbors, distances);
+
+    // Re-map neighbors to their index.
+    for (size_t j = 0; j < neighbors.n_elem; j++)
+      neighbors(j) = indexSame.at(neighbors(j));
+
+    // Store target neighbors.
+    outputMatrix.cols(begin + subIndexSame) = neighbors;
+  }
+}
+
 // Calculates k differently labeled nearest neighbors.
 void Constraints::Impostors(arma::Mat<size_t>& outputMatrix)
 {
@@ -115,15 +156,58 @@ void Constraints::Impostors(arma::Mat<size_t>& outputMatrix)
   }
 }
 
+// Calculates k differently labeled nearest neighbors on a
+// batch of data points.
+void Constraints::Impostors(arma::Mat<size_t>& outputMatrix,
+                            const size_t begin,
+                            const size_t batchSize)
+{
+  arma::mat subDataset = dataset.cols(begin, begin + batchSize - 1);
+  arma::Row<size_t> sublabels = labels.cols(begin, begin + batchSize - 1);
+
+  arma::Row<size_t> uniqueLabels = arma::unique(labels);
+
+  // KNN instance.
+  neighbor::KNN knn;
+
+  arma::Mat<size_t> neighbors;
+  arma::mat distances;
+
+  // Vectors to store indices.
+  arma::uvec indexSame;
+  arma::uvec indexDiff;
+
+  for ( size_t i = 0; i < uniqueLabels.n_cols; i++)
+  {
+    // Calculate impostors.
+    indexSame = arma::find(sublabels == uniqueLabels[i]);
+    indexDiff = arma::find(labels != uniqueLabels[i]);
+
+    // Perform KNN search with differently labeled points as reference
+    // set and same class points as query set.
+    knn.Train(dataset.cols(indexDiff));
+    knn.Search(subDataset.cols(indexSame), k, neighbors, distances);
+
+    // Re-map neighbors to their index.
+    for (size_t j = 0; j < neighbors.n_elem; j++)
+      neighbors(j) = indexDiff.at(neighbors(j));
+
+    // Store impostors.
+    outputMatrix.cols(begin + indexSame) =  neighbors;
+  }
+}
+
[…]
