### 2018-06-02 481fec0
add cosine_search
