+++ src/mlpack/methods/cf/normalization/item_mean_normalization.hpp
@@ -81,8 +81,8 @@ class ItemMeanNormalization
    * @param item Item ID.
    * @param rating Computed rating before denormalization.
    */
-  double Denormalize(const int /* user */,
-                     const int item,
+  double Denormalize(const size_t /* user */,
+                     const size_t item,
                      const double rating) const
   {
     return rating + itemMean(item);
