### 2018-06-03 062c5c6
add some tests
