### 2018-06-04 #726 @mlgill
@Kevin-Mattheus-Moerman

I am still in the process of getting the Python bindings installed. From some of @rasbt's comments on the repo issues, it seems I am not the only one. I will have to work on it later this week.

@rcurtin Thanks for updating the DOIs. Can you rebuild the PDF so I can look at that directly? Thanks.
