### 2018-06-04 3a7bb6a
sampling layer done, kl divergence forward implemented
