### 2018-06-04 c0750ff
kl backward implemented
