### 2018-06-04 e61d021
fix style errors
