### 2018-06-04 c17f481
Fix condition, thanks Kirill for pointing it out.
