### 2018-06-04 cad5a83
Add section on using the Python bindings without installing.
