### 2018-06-04 c4d5744
Clarify any need for LD_LIBRARY_PATH setting.
