### 2018-06-04 657b1a2
Rewrite kRANN documentation to use language-agnostic macros.
