### 2018-06-04 af1f4d1
Point out that Ubuntu repositories may not be up to date.
