### 2018-06-04 mlpack/mlpack#1416 @rcurtin
These should be fixed when #1421 is merged.  Thanks again for pointing the issues out, I really appreciate your dive in.  Many times it can be hard to get good feedback from an outsider's perspective.
