### 2018-06-04 mlpack/mlpack#1418 @rcurtin
Fixed with #1421.
