### 2018-06-04 mlpack/mlpack#1419 @rcurtin
Good call---I have a fix in #1421.  This is one of the tricky bits with writing agnostic documentation: for the command-line, this makes sense.  The model just gets called `perceptron.bin`.  But in Python it goes south fast like you pointed out. :)

I checked all the other bindings too, and none have the same issue.
