### 2018-06-04 mlpack/mlpack#1406 @rcurtin
I added a note to the README and the build guide that will be merged with #1421.
