### 2018-06-04 #726 @rcurtin
Thanks reviewers for your attention to detail.  We've done our best to make the build and installation process as easy as possible, but there is always some level of trickiness at the border of C++ and Python (and I'm sure it will get worse as we add other languages...).

We'll merge mlpack#1421 soon and this should fix the rest of the open issues.
