### 2018-06-04 1cf78f1
suggested changes made
