### 2018-06-04 c1d3eec
changed names in cmakelists
