+++ src/mlpack/methods/ann/layer/layer_types.hpp
@@ -36,7 +36,7 @@
 #include <mlpack/methods/ann/layer/mean_pooling.hpp>
 #include <mlpack/methods/ann/layer/parametric_relu.hpp>
 #include <mlpack/methods/ann/layer/reinforce_normal.hpp>
-#include <mlpack/methods/ann/layer/sampling.hpp>
+#include <mlpack/methods/ann/layer/reparametrization.hpp>
 #include <mlpack/methods/ann/layer/select.hpp>
 #include <mlpack/methods/ann/layer/subview.hpp>
 
@@ -57,7 +57,7 @@ template<typename InputDataType, typename OutputDataType> class LinearNoBias;
 template<typename InputDataType, typename OutputDataType> class LSTM;
 template<typename InputDataType, typename OutputDataType> class GRU;
 template<typename InputDataType, typename OutputDataType> class FastLSTM;
-template<typename InputDataType, typename OutputDataType> class Sampling;
+template<typename InputDataType, typename OutputDataType> class Reparametrization;
 template<typename InputDataType, typename OutputDataType> class VRClassReward;
 
 template<typename InputDataType,
@@ -181,7 +181,7 @@ using LayerTypes = boost::variant<
     Recurrent<arma::mat, arma::mat>*,
     RecurrentAttention<arma::mat, arma::mat>*,
     ReinforceNormal<arma::mat, arma::mat>*,
-    Sampling<arma::mat, arma::mat>*,
+    Reparametrization<arma::mat, arma::mat>*,
     Select<arma::mat, arma::mat>*,
     Sequential<arma::mat, arma::mat>*,
     Subview<arma::mat, arma::mat>*,
