### 2018-06-05 11bfec2
corrected kl forward and backward
