+++ src/mlpack/tests/lmnn_test.cpp
@@ -145,10 +145,10 @@ BOOST_AUTO_TEST_CASE(LMNNInitialGradient)
   lmnnfn.Gradient(coordinates, gradient);
 
   // Result calculated by hand.
-  BOOST_REQUIRE_CLOSE(gradient(0, 0), -0.144, 1e-5);
+  BOOST_REQUIRE_CLOSE(gradient(0, 0), -0.288, 1e-5);
   BOOST_REQUIRE_SMALL(gradient(1, 0), 1e-5);
   BOOST_REQUIRE_SMALL(gradient(0, 1), 1e-5);
-  BOOST_REQUIRE_CLOSE(gradient(1, 1), 6.0, 1e-5);
+  BOOST_REQUIRE_CLOSE(gradient(1, 1), 12.0, 1e-5);
 }
 
 /***
@@ -170,10 +170,10 @@ BOOST_AUTO_TEST_CASE(LMNNInitialEvaluateWithGradient)
   // Result calculated by hand.
   BOOST_REQUIRE_CLOSE(objective, 9.456, 1e-5);
   // Check Gradient
-  BOOST_REQUIRE_CLOSE(gradient(0, 0), -0.144, 1e-5);
+  BOOST_REQUIRE_CLOSE(gradient(0, 0), -0.288, 1e-5);
   BOOST_REQUIRE_SMALL(gradient(1, 0), 1e-5);
   BOOST_REQUIRE_SMALL(gradient(0, 1), 1e-5);
-  BOOST_REQUIRE_CLOSE(gradient(1, 1), 6.0, 1e-5);
+  BOOST_REQUIRE_CLOSE(gradient(1, 1), 12.0, 1e-5);
 }
 
 /**
@@ -215,45 +215,45 @@ BOOST_AUTO_TEST_CASE(LMNNSeparableGradient)
 
   lmnnfn.Gradient(coordinates, 0, gradient, 1);
 
-  BOOST_REQUIRE_CLOSE(gradient(0, 0), -0.024, 1e-5);
+  BOOST_REQUIRE_CLOSE(gradient(0, 0), -0.048, 1e-5);
   BOOST_REQUIRE_CLOSE(gradient(0, 1), 0.0, 1e-5);
   BOOST_REQUIRE_CLOSE(gradient(1, 0), 0.0, 1e-5);
-  BOOST_REQUIRE_CLOSE(gradient(1, 1), 1.0, 1e-5);
+  BOOST_REQUIRE_CLOSE(gradient(1, 1), 2.0, 1e-5);
 
   lmnnfn.Gradient(coordinates, 1, gradient, 1);
 
-  BOOST_REQUIRE_CLOSE(gradient(0, 0), -0.024, 1e-5);
+  BOOST_REQUIRE_CLOSE(gradient(0, 0), -0.048, 1e-5);
   BOOST_REQUIRE_CLOSE(gradient(0, 1), 0.0, 1e-5);
   BOOST_REQUIRE_CLOSE(gradient(1, 0), 0.0, 1e-5);
-  BOOST_REQUIRE_CLOSE(gradient(1, 1), 1.0, 1e-5);
+  BOOST_REQUIRE_CLOSE(gradient(1, 1), 2.0, 1e-5);
 
   lmnnfn.Gradient(coordinates, 2, gradient, 1);
 
-  BOOST_REQUIRE_CLOSE(gradient(0, 0), -0.024, 1e-5);
+  BOOST_REQUIRE_CLOSE(gradient(0, 0), -0.048, 1e-5);
   BOOST_REQUIRE_CLOSE(gradient(0, 1), 0.0, 1e-5);
   BOOST_REQUIRE_CLOSE(gradient(1, 0), 0.0, 1e-5);
[…]
