### 2018-06-06 c614806
Remove warnings and added some more tests
