+++ src/mlpack/tests/lmnn_test.cpp
@@ -1,6 +1,8 @@
 /**
  * @file lmnn_test.cpp
+ * @author Ryan Curtin
  * @author Manish Kumar
+ *
  * Unit tests for Large Margin Nearest Neighbors and related code (including
  * the constraints class).
  *
@@ -13,6 +15,7 @@
 #include <mlpack/core/metrics/lmetric.hpp>
 #include <mlpack/methods/lmnn/lmnn.hpp>
 #include <mlpack/core/optimizers/lbfgs/lbfgs.hpp>
+#include <mlpack/methods/neighbor_search/neighbor_search.hpp>
 
 #include <boost/test/unit_test.hpp>
 #include "test_tools.hpp"
@@ -113,7 +116,7 @@ BOOST_AUTO_TEST_CASE(LMNNInitialPointTest)
 /***
  * Ensure non-seprable objective function is right.
  */
-BOOST_AUTO_TEST_CASE(LMNNInitialEvaluation)
+BOOST_AUTO_TEST_CASE(LMNNInitialEvaluationTest)
 {
   // Useful but simple dataset with six points and two classes.
   arma::mat dataset        = "-0.1 -0.1 -0.1  0.1  0.1  0.1;"
@@ -131,7 +134,7 @@ BOOST_AUTO_TEST_CASE(LMNNInitialEvaluation)
 /**
  * Ensure non-seprable gradient function is right.
  */
-BOOST_AUTO_TEST_CASE(LMNNInitialGradient)
+BOOST_AUTO_TEST_CASE(LMNNInitialGradientTest)
 {
   // Useful but simple dataset with six points and two classes.
   arma::mat dataset        = "-0.1 -0.1 -0.1  0.1  0.1  0.1;"
@@ -154,7 +157,7 @@ BOOST_AUTO_TEST_CASE(LMNNInitialGradient)
 /***
  * Ensure non-seprable EvaluateWithGradient function is right.
  */
-BOOST_AUTO_TEST_CASE(LMNNInitialEvaluateWithGradient)
+BOOST_AUTO_TEST_CASE(LMNNInitialEvaluateWithGradientTest)
 {
   // Useful but simple dataset with six points and two classes.
   arma::mat dataset        = "-0.1 -0.1 -0.1  0.1  0.1  0.1;"
@@ -179,7 +182,7 @@ BOOST_AUTO_TEST_CASE(LMNNInitialEvaluateWithGradient)
 /**
  * Ensure the separable objective function is right.
  */
-BOOST_AUTO_TEST_CASE(LMNNSeparableObjective)
+BOOST_AUTO_TEST_CASE(LMNNSeparableObjectiveTest)
 {
   // Useful but simple dataset with six points and two classes.
   arma::mat dataset        = "-0.1 -0.1 -0.1  0.1  0.1  0.1;"
@@ -201,7 +204,7 @@ BOOST_AUTO_TEST_CASE(LMNNSeparableObjective)
 /**
  * Ensure the separable gradient is right.
  */
-BOOST_AUTO_TEST_CASE(LMNNSeparableGradient)
+BOOST_AUTO_TEST_CASE(LMNNSeparableGradientTest)
 {
   // Useful but simple dataset with six points and two classes.
   arma::mat dataset           = "-0.1 -0.1 -0.1  0.1  0.1  0.1;"
@@ -259,7 +262,7 @@ BOOST_AUTO_TEST_CASE(LMNNSeparableGradient)
 /**
  * Ensure the separable EvaluateWithGradient function is right.
  */
-BOOST_AUTO_TEST_CASE(LMNNSeparableEvaluateWithGradient)
+BOOST_AUTO_TEST_CASE(LMNNSeparableEvaluateWithGradientTest)
 {
[…]
