+++ src/mlpack/methods/lmnn/lmnn_function_impl.hpp
@@ -176,8 +176,8 @@ void LMNNFunction<MetricType>::Gradient(const arma::mat& coordinates,
 {
   gradient.zeros(coordinates.n_rows, coordinates.n_cols);
 
-  arma::mat cij = arma::zeros(coordinates.n_rows, coordinates.n_cols);
-  arma::mat cil = arma::zeros(coordinates.n_rows, coordinates.n_cols);
+  arma::mat cij = arma::zeros(dataset.n_rows, dataset.n_rows);
+  arma::mat cil = arma::zeros(dataset.n_rows, dataset.n_rows);
 
   for (size_t i = 0; i < dataset.n_cols; i++)
   {
@@ -230,8 +230,8 @@ void LMNNFunction<MetricType>::Gradient(const arma::mat& coordinates,
 {
   gradient.zeros(coordinates.n_rows, coordinates.n_cols);
 
-  arma::mat cij = arma::zeros(coordinates.n_rows, coordinates.n_cols);
-  arma::mat cil = arma::zeros(coordinates.n_rows, coordinates.n_cols);
+  arma::mat cij = arma::zeros(dataset.n_rows, dataset.n_rows);
+  arma::mat cil = arma::zeros(dataset.n_rows, dataset.n_rows);
 
   for (size_t i = begin; i < begin + batchSize; i++)
   {
@@ -292,8 +292,8 @@ double LMNNFunction<MetricType>::EvaluateWithGradient(
 
   gradient.zeros(coordinates.n_rows, coordinates.n_cols);
 
-  arma::mat cij = arma::zeros(coordinates.n_rows, coordinates.n_cols);
-  arma::mat cil = arma::zeros(coordinates.n_rows, coordinates.n_cols);
+  arma::mat cij = arma::zeros(dataset.n_rows, dataset.n_rows);
+  arma::mat cil = arma::zeros(dataset.n_rows, dataset.n_rows);
 
   for (size_t i = 0; i < dataset.n_cols; i++)
   {
@@ -366,8 +366,8 @@ double LMNNFunction<MetricType>::EvaluateWithGradient(
 
   gradient.zeros(coordinates.n_rows, coordinates.n_cols);
 
-  arma::mat cij = arma::zeros(coordinates.n_rows, coordinates.n_cols);
-  arma::mat cil = arma::zeros(coordinates.n_rows, coordinates.n_cols);
+  arma::mat cij = arma::zeros(dataset.n_rows, dataset.n_rows);
+  arma::mat cil = arma::zeros(dataset.n_rows, dataset.n_rows);
 
   for (size_t i = begin; i < begin + batchSize; i++)
   {
