### 2018-06-06 e09121b
Precalculate gradient due to target neighbors
