+++ src/mlpack/methods/lmnn/lmnn_function_impl.hpp
@@ -31,7 +31,9 @@ LMNNFunction<MetricType>::LMNNFunction(const arma::mat& dataset,
     labels(math::MakeAlias(const_cast<arma::Row<size_t>&>(labels), false)),
     k(k),
     metric(metric),
-    regularization(regularization)
+    regularization(regularization),
+    p_cij(dataset.n_cols),
+    precalculated(false)
 {
   // Initialize the initial learning point.
   initialPoint.eye(dataset.n_rows, dataset.n_rows);
@@ -179,13 +181,15 @@ void LMNNFunction<MetricType>::Gradient(const arma::mat& coordinates,
   arma::mat cij = arma::zeros(dataset.n_rows, dataset.n_rows);
   arma::mat cil = arma::zeros(dataset.n_rows, dataset.n_rows);
 
+  // Calculate the target neighbors gradient components, if necessary.
+  Precalculate();
+
   for (size_t i = 0; i < dataset.n_cols; i++)
   {
     for (size_t j = 0; j < k ; j++)
     {
       // Calculate gradient due to target neighbors.
-      arma::vec diff = dataset.col(i) - dataset.col(targetNeighbors(j, i));
-      cij += diff * arma::trans(diff);
+      cij += p_cij[i][j];
     }
 
     for (int j = k - 1; j >= 0; j--)
@@ -207,10 +211,9 @@ void LMNNFunction<MetricType>::Gradient(const arma::mat& coordinates,
           break;
         }
 
-        arma::vec diff = dataset.col(i) - dataset.col(targetNeighbors(j, i));
-        cil += diff * arma::trans(diff);
+        cil += p_cij[i][j];
 
-        diff = dataset.col(i) - dataset.col(impostors(l, i));
+        arma::vec diff = dataset.col(i) - dataset.col(impostors(l, i));
         cil -= diff * arma::trans(diff);
       }
     }
@@ -233,13 +236,15 @@ void LMNNFunction<MetricType>::Gradient(const arma::mat& coordinates,
   arma::mat cij = arma::zeros(dataset.n_rows, dataset.n_rows);
   arma::mat cil = arma::zeros(dataset.n_rows, dataset.n_rows);
 
+  // Calculate the target neighbors gradient components, if necessary.
+  Precalculate();
+
   for (size_t i = begin; i < begin + batchSize; i++)
   {
     for (size_t j = 0; j < k ; j++)
     {
       // Calculate gradient due to target neighbors.
-      arma::vec diff = dataset.col(i) - dataset.col(targetNeighbors(j, i));
-      cij += diff * arma::trans(diff);
+      cij += p_cij[i][j];
     }
 
     for (int j = k - 1; j >= 0; j--)
@@ -261,10 +266,9 @@ void LMNNFunction<MetricType>::Gradient(const arma::mat& coordinates,
           break;
         }
 
-        arma::vec diff = dataset.col(i) - dataset.col(targetNeighbors(j, i));
-        cil += diff * arma::trans(diff);
+        cil += p_cij[i][j];
 
-        diff = dataset.col(i) - dataset.col(impostors(l, i));
+        arma::vec diff = dataset.col(i) - dataset.col(impostors(l, i));
         cil -= diff * arma::trans(diff);
       }
     }
@@ -295,6 +299,9 @@ double LMNNFunction<MetricType>::EvaluateWithGradient(
   arma::mat cij = arma::zeros(dataset.n_rows, dataset.n_rows);
   arma::mat cil = arma::zeros(dataset.n_rows, dataset.n_rows);
 
+  // Calculate the target neighbors gradient components, if necessary.
+  Precalculate();
+
   for (size_t i = 0; i < dataset.n_cols; i++)
   {
[…]
