### 2018-06-06 03e7b1c
change totalError to rmse
