### 2018-06-07 8902358
Remove precalculation---not good for larger datasets
