### 2018-06-07 9d58667
added numerical gradient test
