+++ src/mlpack/tests/ann_layer_test.cpp
@@ -229,7 +229,7 @@ BOOST_AUTO_TEST_CASE(JacobianAddLayerTest)
 }
 
 /**
- * Add layer numerically gradient test.
+ * Add layer numerical gradient test.
  */
 BOOST_AUTO_TEST_CASE(GradientAddLayerTest)
 {
@@ -256,7 +256,7 @@ BOOST_AUTO_TEST_CASE(GradientAddLayerTest)
 
     double Gradient(arma::mat& gradient) const
     {
-      arma::mat output;
+      // arma::mat output;
       double error = model->Evaluate(model->Parameters(), 0, 1);
       model->Gradient(model->Parameters(), 0, gradient, 1);
       return error;
@@ -535,7 +535,7 @@ BOOST_AUTO_TEST_CASE(JacobianLinearLayerTest)
 }
 
 /**
- * Linear layer numerically gradient test.
+ * Linear layer numerical gradient test.
  */
 BOOST_AUTO_TEST_CASE(GradientLinearLayerTest)
 {
@@ -562,7 +562,7 @@ BOOST_AUTO_TEST_CASE(GradientLinearLayerTest)
 
     double Gradient(arma::mat& gradient) const
     {
-      arma::mat output;
+      // arma::mat output;
       double error = model->Evaluate(model->Parameters(), 0, 1);
       model->Gradient(model->Parameters(), 0, gradient, 1);
       return error;
@@ -619,7 +619,7 @@ BOOST_AUTO_TEST_CASE(JacobianLinearNoBiasLayerTest)
 }
 
 /**
- * LinearNoBias layer numerically gradient test.
+ * LinearNoBias layer numerical gradient test.
  */
 BOOST_AUTO_TEST_CASE(GradientLinearNoBiasLayerTest)
 {
@@ -646,7 +646,7 @@ BOOST_AUTO_TEST_CASE(GradientLinearNoBiasLayerTest)
 
     double Gradient(arma::mat& gradient) const
     {
-      arma::mat output;
+      // arma::mat output;
       double error = model->Evaluate(model->Parameters(), 0, 1);
       model->Gradient(model->Parameters(), 0, gradient, 1);
       return error;
@@ -721,7 +721,7 @@ BOOST_AUTO_TEST_CASE(JacobianFlexibleReLULayerTest)
 }
 
 /**
- * Flexible ReLU layer numerically gradient test.
+ * Flexible ReLU layer numerical gradient test.
  */
 BOOST_AUTO_TEST_CASE(GradientFlexibleReLULayerTest)
 {
@@ -750,7 +750,7 @@ BOOST_AUTO_TEST_CASE(GradientFlexibleReLULayerTest)
 
     double Gradient(arma::mat& gradient) const
     {
-      arma::mat output;
+      // arma::mat output;
       double error = model->Evaluate(model->Parameters(), 0, 1);
       model->Gradient(model->Parameters(), 0, gradient, 1);
       return error;
@@ -957,7 +957,7 @@ BOOST_AUTO_TEST_CASE(GradientLSTMLayerTest)
 
     double Gradient(arma::mat& gradient) const
     {
-      arma::mat output;
+      // arma::mat output;
       double error = model->Evaluate(model->Parameters(), 0, 1);
       model->Gradient(model->Parameters(), 0, gradient, 1);
       return error;
@@ -1039,7 +1039,7 @@ BOOST_AUTO_TEST_CASE(GradientFastLSTMLayerTest)
 
     double Gradient(arma::mat& gradient) const
     {
-      arma::mat output;
+      // arma::mat output;
       double error = model->Evaluate(model->Parameters(), 0, 1);
       model->Gradient(model->Parameters(), 0, gradient, 1);
[…]
