### 2018-06-08 309f642
gradient check passed, removed redundant lines
