### 2018-06-08 c20850c
Introduce run parameter to forward the Forward/Backward/Gradient call inside the merge layer.
