### 2018-06-08 3e6b64e
Add merge layer run test.
