### 2018-06-08 f96be8b
set zero rating to small double
