+++ src/mlpack/methods/cf/normalization/item_mean_normalization.hpp
@@ -56,6 +56,10 @@ class ItemMeanNormalization
     data.each_col([&](arma::vec& datapoint) {
       const size_t item = (size_t) datapoint(1);
       datapoint(2) -= itemMean(item);
+      // The algorithm omits rating of zero. If normalized rating equals zero,
+      // it is set to the smallest positive double value.
+      if (datapoint(2) == 0)
+        datapoint(2) = std::numeric_limits<double>::min();
     });
   }
 
@@ -71,7 +75,13 @@ class ItemMeanNormalization
     arma::sp_mat::iterator it = cleanedData.begin();
     arma::sp_mat::iterator it_end = cleanedData.end();
     for (; it != it_end; it++)
+    {
       *it = *it - itemMean(it.row());
+      // The algorithm omits rating of zero. If normalized rating equals zero,
+      // it is set to the smallest positive double value.
+      if (*it == 0)
+        *it = std::numeric_limits<double>::min();
+    }
   }
 
   /**
