### 2018-06-08 mlpack/mlpack#1416 @rcurtin
Fixed via #1421.
