### 2018-06-08 mlpack/mlpack#1418 @rcurtin
#1421 is merged, so I will close this now.
