### 2018-06-08 mlpack/mlpack#1419 @rcurtin
Fixed with the merge of #1421.
