### 2018-06-08 mlpack/mlpack#1406 @rcurtin
Since #1421 is merged, I'll go ahead and close this now.
