### 2018-06-08 88ae069
corrections made in kl, more tests added
