### 2018-06-08 ad90acb
Fix minor style issues (remove blank line, add line at the end of the file).
