### 2018-06-09 a33149f
replace() not supported
