### 2018-06-09 7c6c8a4
add precalculation for constraints
