### 2018-06-09 66faebc
modify rmse bound in cf_test
