### 2018-06-09 061d86e
add tests for normalization
