### 2018-06-09 dcfef4f
loss visitor added
