### 2018-06-09 65c8015
loss visitor added
