### 2018-06-09 #726 @rasbt
@Kevin-Mattheus-Moerman

happy to let you know that all my issues were addressed in the recent merge of PR [mlpack/mlpack#1421](https://github.com/mlpack/mlpack/pull/1421). Aside from bumping the version to 3.0.2 in the paper and linking to the 3.0.2 archive (I think Joss uses Zenodo?; I think this should probably be done after the other reviewers finished the review, though), I am happy to approve this submission :)
