### 2018-06-10 c220d06
add batch targetNeighbors gradient precalculation
