### 2018-06-10 2289d1a
complete RegressionInterpolation class
