### 2018-06-10 1959b50
Use the correct batch size.
