### 2018-06-10 #726 @mlgill
@Kevin-Mattheus-Moerman It looks like someone may have marked the installation issues under my name by mistake, but I can confirm that those items are complete. As @rasbt mentioned, the version bump is the only thing remaining. Thanks!
