### 2018-06-11 86f4fc9
Constructor edit
