### 2018-06-11 1aef628
Constructor default params
