### 2018-06-11 b965b07
update comments & debug
