+++ src/mlpack/methods/cf/normalization/user_mean_normalization.hpp
@@ -32,6 +32,7 @@ namespace cf {
  *
  * // Generate 10 recommendations for all users.
  * cf.GetRecommendations(10, recommendations);
+ * @endcode
  */
 class UserMeanNormalization
 {
