+++ src/mlpack/methods/cf/normalization/item_mean_normalization.hpp
@@ -147,7 +147,7 @@ class ItemMeanNormalization
   }
 
  private:
-  //! item mean.
+  //! Item mean.
   arma::vec itemMean;
 };
 
