+++ src/mlpack/methods/cf/normalization/z_score_normalization.hpp
@@ -31,6 +31,7 @@ namespace cf {
  *
  * // Generate 10 recommendations for all users.
  * cf.GetRecommendations(10, recommendations);
+ * @endcode
  */
 class ZScoreNormalization
 {
