### 2018-06-11 beaa319
remove if(cleanData) block
