+++ src/mlpack/tests/cf_test.cpp
@@ -42,7 +42,7 @@ using namespace std;
  * set. Default case.
  */
 template<typename DecompositionPolicy>
-void GetRecommendationsAllUsers(bool cleanData = true)
+void GetRecommendationsAllUsers()
 {
   DecompositionPolicy decomposition;
   // Dummy number of recommendations.
@@ -57,15 +57,6 @@ void GetRecommendationsAllUsers(bool cleanData = true)
   arma::mat dataset;
   data::Load("GroupLensSmall.csv", dataset);
 
-  if (cleanData)
-  {
-    // Make data into sparse matrix.
-    arma::sp_mat cleanedData;
-    CFType<>::CleanData(dataset, cleanedData);
-
-    arma::sp_mat dataset;
-    dataset = cleanedData;
-  }
   CFType<> c(dataset, decomposition, 5, 5, 70);
 
   // Generate recommendations when query set is not specified.
@@ -82,7 +73,7 @@ void GetRecommendationsAllUsers(bool cleanData = true)
  * Make sure that the recommendations are generated for queried users only.
  */
 template<typename DecompositionPolicy>
-void GetRecommendationsQueriedUser(bool cleanData = true)
+void GetRecommendationsQueriedUser()
 {
   DecompositionPolicy decomposition;
   // Number of users that we will search for recommendations for.
@@ -103,16 +94,6 @@ void GetRecommendationsQueriedUser(bool cleanData = true)
   arma::mat dataset;
   data::Load("GroupLensSmall.csv", dataset);
 
-  if (cleanData)
-  {
-    // Make data into sparse matrix.
-    arma::sp_mat cleanedData;
-    CFType<>::CleanData(dataset, cleanedData);
-
-    arma::sp_mat dataset;
-    dataset = cleanedData;
-  }
-
   CFType<> c(dataset, decomposition, 5, 5, 70);
 
   // Generate recommendations when query set is specified.
@@ -130,7 +111,7 @@ void GetRecommendationsQueriedUser(bool cleanData = true)
  */
 template<typename DecompositionPolicy,
          typename NormalizationType = NoNormalization>
-void RecommendationAccuracy(bool cleanData = true)
+void RecommendationAccuracy()
 {
   DecompositionPolicy decomposition;
   // Load the GroupLens dataset; then, we will remove some values from it.
@@ -169,16 +150,6 @@ void RecommendationAccuracy(bool cleanData = true)
         ++currentCol;
       }
     }
-  }
-
-  if (cleanData)
-  {
-    // Make data into sparse matrix.
-    arma::sp_mat cleanedData;
-    CFType<NormalizationType>::CleanData(dataset, cleanedData);
-
-    arma::sp_mat dataset;
-    dataset = cleanedData;
   }
 
   CFType<NormalizationType> c(dataset, decomposition, 5, 5, 70);
@@ -232,7 +203,7 @@ void RecommendationAccuracy(bool cleanData = true)
 // Make sure that Predict() is returning reasonable results.
 template<typename DecompositionPolicy,
          typename NormalizationType = NoNormalization>
-void CFPredict(bool cleanData = true, const double rmseBound = 2.0)
+void CFPredict(const double rmseBound = 2.0)
 {
   DecompositionPolicy decomposition;
[…]
