### 2018-06-11 312b155
extra loss added to evaluate, backward
