+++ src/mlpack/methods/ann/ffn.hpp
@@ -23,6 +23,7 @@
 #include "visitor/reset_visitor.hpp"
 #include "visitor/weight_size_visitor.hpp"
 #include "visitor/copy_visitor.hpp"
+#include "visitor/loss_visitor.hpp"
 
 #include "init_rules/network_init.hpp"
 
@@ -370,6 +371,9 @@ class FFN
 
   //! Locally-stored output height visitor.
   OutputHeightVisitor outputHeightVisitor;
+
+  //! Locally-stored loss visitor
+  LossVisitor lossVisitor;
 
   //! Locally-stored reset visitor.
   ResetVisitor resetVisitor;
