### 2018-06-11 3dbbe13
removed simple kl test
