### 2018-06-11 059fe89
removed simple kl test
