### 2018-06-12 5d9992d
advanced ctor
