### 2018-06-12 8502230
use complete sentences for examples
