+++ src/mlpack/methods/cf/normalization/user_mean_normalization.hpp
@@ -24,10 +24,12 @@ namespace cf {
  * An example of how to use UserMeanNormalization in CF is shown below:
  *
  * @code
- * extern arma::mat data; // (user, item, rating) table
- * extern arma::Col<size_t> users; // users seeking recommendations
- * arma::Mat<size_t> recommendations; // Recommendations
+ * extern arma::mat data; // data is a (user, item, rating) table.
+ * // Users for whom recommendations are generated.
+ * extern arma::Col<size_t> users;
+ * arma::Mat<size_t> recommendations; // Resulting recommendations.
  *
+ * // Use UserMeanNormalization as normalization method.
  * CFType<UserMeanNormalization> cf(data);
  *
  * // Generate 10 recommendations for all users.
