### 2018-06-12 de67371
change method names to Mean() and return const refenrence
