+++ src/mlpack/methods/cf/normalization/user_mean_normalization.hpp
@@ -134,10 +134,7 @@ class UserMeanNormalization
   /**
    * Return user mean.
    */
-  arma::vec UserMean() const
-  {
-    return userMean;
-  }
+  const arma::vec& Mean() const { return userMean; }
 
   /**
    * Serialization.
