+++ src/mlpack/methods/cf/normalization/user_mean_normalization.hpp
@@ -83,7 +83,7 @@ class UserMeanNormalization
   /**
    * Normalize the data by subtracting user mean from each of existing rating.
    *
-   * @param cleanedData Sparse matrix data.
+   * @param cleanedData Input data as a sparse matrix.
    */
   void Normalize(arma::sp_mat& cleanedData)
   {
