+++ src/mlpack/methods/cf/normalization/overall_mean_normalization.hpp
@@ -62,7 +62,7 @@ class OverallMeanNormalization
   /**
    * Normalize the data by subtracting the mean of all existing ratings.
    *
-   * @param cleanedData Sparse matrix data.
+   * @param cleanedData Input data as a sparse matrix.
    */
   void Normalize(arma::sp_mat& cleanedData)
   {
