+++ src/mlpack/methods/cf/normalization/combined_normalization.hpp
@@ -61,7 +61,7 @@ class CombinedNormalization
   /**
    * Normalize the data by calling Normalize() in each normalization object.
    *
-   * @param cleanedData Sparse matrix data.
+   * @param cleanedData Input data as a sparse matrix.
    */
   void Normalize(arma::sp_mat& cleanedData)
   {
