### 2018-06-12 98a1f64
templatize Normalize() functions in some classes
