### 2018-06-12 ab0f62d
Fix bibtex citation.
