### 2018-06-12 342972c
Implement Batch Support
