+++ src/mlpack/methods/ann/layer/transposed_convolution_impl.hpp
@@ -114,30 +114,39 @@ void TransposedConvolution<
     OutputDataType
 >::Forward(const arma::Mat<eT>&& input, arma::Mat<eT>&& output)
 {
-  inputTemp = arma::cube(input.memptr(), inputWidth, inputHeight, inSize);
+  batchSize = input.n_cols;
+  inputTemp = arma::cube(const_cast<arma::Mat<eT>&&>(input).memptr(),
+      inputWidth, inputHeight, inSize * batchSize, false, false);
 
   outputWidth = TransposedConvOutSize(inputWidth, kW, dW, padW);
   outputHeight = TransposedConvOutSize(inputHeight, kH, dH, padH);
 
-  output.set_size(outputWidth * outputHeight * outSize, 1);
+  output.set_size(outputWidth * outputHeight * outSize, batchSize);
   outputTemp = arma::Cube<eT>(output.memptr(), outputWidth, outputHeight,
-      outSize, false, false);
+      outSize * batchSize, false, false);
   outputTemp.zeros();
 
-  for (size_t outMap = 0, outMapIdx = 0; outMap < outSize; outMap++)
+  for (size_t outMap = 0, outMapIdx = 0, batchCount = 0; outMap <
+      outSize * batchSize; outMap++)
   {
+    if (outMap != 0 && outMap % outSize == 0)
+    {
+      batchCount++;
+      outMapIdx = 0;
+    }
+
     for (size_t inMap = 0; inMap < inSize; inMap++, outMapIdx++)
     {
       arma::Mat<eT> convOutput, rotatedFilter;
       Rotate180(weight.slice(outMapIdx), rotatedFilter);
 
-      BackwardConvolutionRule::Convolution(inputTemp.slice(inMap),
-          rotatedFilter, convOutput, 1, 1);
+      BackwardConvolutionRule::Convolution(inputTemp.slice(inMap +
+          batchCount * inSize), rotatedFilter, convOutput, 1, 1);
 
       outputTemp.slice(outMap) += convOutput;
     }
 
-    outputTemp.slice(outMap) += bias(outMap);
+    outputTemp.slice(outMap) += bias(outMap % outSize);
   }
 }
 
@@ -158,16 +167,23 @@ void TransposedConvolution<
 >::Backward(
     const arma::Mat<eT>&& /* input */, arma::Mat<eT>&& gy, arma::Mat<eT>&& g)
 {
-  arma::cube mappedError(gy.memptr(), outputWidth, outputHeight, outSize,
-      false, false);
-
-  g.set_size(inputTemp.n_rows * inputTemp.n_cols * inputTemp.n_slices, 1);
+  arma::cube mappedError(gy.memptr(), outputWidth, outputHeight,
+      outSize * batchSize, false, false);
+  g.set_size(inputTemp.n_rows * inputTemp.n_cols * inSize, batchSize);
   gTemp = arma::Cube<eT>(g.memptr(), inputTemp.n_rows,
       inputTemp.n_cols, inputTemp.n_slices, false, false);
+
   gTemp.zeros();
 
-  for (size_t outMap = 0, outMapIdx = 0; outMap < outSize; outMap++)
+  for (size_t outMap = 0, outMapIdx = 0, batchCount = 0; outMap <
+      outSize * batchSize; outMap++)
   {
+    if (outMap != 0 && outMap % outSize == 0)
+    {
+      batchCount++;
+      outMapIdx = 0;
+    }
+
     for (size_t inMap = 0; inMap < inSize; inMap++, outMapIdx++)
     {
       arma::Mat<eT> output;
@@ -175,7 +191,7 @@ void TransposedConvolution<
[…]
