+++ src/mlpack/methods/ann/layer/atrous_convolution_impl.hpp
@@ -118,7 +118,9 @@ void AtrousConvolution<
     OutputDataType
 >::Forward(const arma::Mat<eT>&& input, arma::Mat<eT>&& output)
 {
-  inputTemp = arma::cube(input.memptr(), inputWidth, inputHeight, inSize);
+  batchSize = input.n_cols;
+  inputTemp = arma::cube(const_cast<arma::Mat<eT>&&>(input).memptr(),
+      inputWidth, inputHeight, inSize * batchSize, false, false);
 
   if (padW != 0 || padH != 0)
   {
@@ -128,30 +130,41 @@ void AtrousConvolution<
   size_t wConv = ConvOutSize(inputWidth, kW, dW, padW, dilationW);
   size_t hConv = ConvOutSize(inputHeight, kH, dH, padH, dilationH);
 
-  output.set_size(wConv * hConv * outSize, 1);
-  outputTemp = arma::Cube<eT>(output.memptr(), wConv, hConv, outSize,
-      false, false);
+  output.set_size(wConv * hConv * outSize, batchSize);
+  outputTemp = arma::Cube<eT>(output.memptr(), wConv, hConv,
+      outSize * batchSize, false, false);
   outputTemp.zeros();
 
-  for (size_t outMap = 0, outMapIdx = 0; outMap < outSize; outMap++)
+  for (size_t outMap = 0, outMapIdx = 0, batchCount = 0; outMap <
+      outSize * batchSize; outMap++)
   {
+    if (outMap != 0 && outMap % outSize == 0)
+    {
+      batchCount++;
+      outMapIdx = 0;
+    }
+
     for (size_t inMap = 0; inMap < inSize; inMap++, outMapIdx++)
     {
       arma::Mat<eT> convOutput;
+
       if (padW != 0 || padH != 0)
       {
-        ForwardConvolutionRule::Convolution(inputPaddedTemp.slice(inMap),
-            weight.slice(outMapIdx), convOutput, dW, dH, dilationW, dilationH);
+        ForwardConvolutionRule::Convolution(inputPaddedTemp.slice(inMap +
+            batchCount * inSize), weight.slice(outMapIdx), convOutput, dW, dH,
+            dilationW, dilationH);
       }
       else
       {
-        ForwardConvolutionRule::Convolution(inputTemp.slice(inMap),
-            weight.slice(outMapIdx), convOutput, dW, dH, dilationW, dilationH);
+        ForwardConvolutionRule::Convolution(inputTemp.slice(inMap +
+            batchCount * inSize), weight.slice(outMapIdx), convOutput, dW, dH,
+            dilationW, dilationH);
       }
+
       outputTemp.slice(outMap) += convOutput;
     }
 
-    outputTemp.slice(outMap) += bias(outMap);
+    outputTemp.slice(outMap) += bias(outMap % outSize);
   }
 
   outputWidth = outputTemp.n_rows;
@@ -175,16 +188,23 @@ void AtrousConvolution<
 >::Backward(
     const arma::Mat<eT>&& /* input */, arma::Mat<eT>&& gy, arma::Mat<eT>&& g)
 {
-  arma::cube mappedError(gy.memptr(), outputWidth, outputHeight, outSize,
-      false, false);
+  arma::cube mappedError(gy.memptr(), outputWidth, outputHeight,
+      outSize * batchSize, false, false);
 
-  g.set_size(inputTemp.n_rows * inputTemp.n_cols * inputTemp.n_slices, 1);
+  g.set_size(inputTemp.n_rows * inputTemp.n_cols * inSize, batchSize);
   gTemp = arma::Cube<eT>(g.memptr(), inputTemp.n_rows,
       inputTemp.n_cols, inputTemp.n_slices, false, false);
[…]
