### 2018-06-12 be2869d
Shuffle the entire data matrix instead of keeping a visitationOrder.
