### 2018-06-13 b65a22a
windows build tutorial using up-to-date versions
