+++ doc/guide/build_windows.hpp
@@ -1,0 +1,95 @@
+/*! @page build_windows Building mlpack From Source
+
+@section build_intro Introduction
+
+This document discusses how to build mlpack for Windows from source, so you can later
+create your own C++ applications.
+
+@section build_env Environment
+
+This tutorial has been designed and tested using:
+- Windows 10
+- Visual Studio 2017 (toolset v141)
+- mlpack-3.0.2
+- OpenBLAS.0.2.14.1
+- boost_1_66_0-msvc-14.1-64
+- armadillo-8.500.1
+- and x64 configuration
+
+The directories and paths used in this tutorial are just for reference purposes.
+
+@section build_prereqs Pre-requisites
+
+- Install CMake for Windows (win64-x64 version from https://cmake.org/download/)
+and make sure you can use it from the Command Prompt (may need to add to the PATH)
+
+- Download the latest mlpack release from here:
+<a href="http://www.mlpack.org/files/mlpack-3.0.2.tar.gz">mlpack-3.0.2</a>
+
+@section build_instructions Windows build instructions
+
+- Unzip mlpack to "C:\mlpack\mlpack-3.0.2"
+- Open Visual Studio and select: File > New > Project from Existing Code
+ - Type of project: Visual C++
+ - Project location: "C:\mlpack\mlpack-3.0.2"
+ - Project name: mlpack
+ - Finish
+- We will use this Visual Studio project to get the OpenBLAS dependency in the next section
+
+@section build_dependencies Dependencies
+
+<b> OpenBLAS Dependency </b>
+
+- Open the NuGet packages manager (Tools > NuGet Package Manager > Manage NuGet Packages for Solution...)
+- Click on the “Browse” tab and search for “openblas”
+- Click on OpenBlas and check the mlpack project, then click Install
+- Once it has finished installing, close Visual Studio
+
+<b> Boost Dependency </b>
+
+You can either get Boost via NuGet or you can download the prebuilt Windows binaries separately.
+This tutorial follows the second approach for simplicity.
+
+- Download the "Prebuilt Windows binaries" of the Boost library ("boost_1_66_0-msvc-14.1-64") from 
+<a href="https://sourceforge.net/projects/boost/files/boost-binaries/">Sourceforge</a>
+
+@note Make sure you download the MSVC version that matches your Visual Studio
+
+- Install or unzip to "C:\boost\boost_1_66_0"
+
+<b> Armadillo Dependency </b>
+
+- Download "Armadillo" (armadillo-8.500.1.tar.xz) from <a href="http://arma.sourceforge.net/download.html">Sourceforge</a>
+- Unzip to "C:\mlpack\armadillo-8.500.1"
+- Create a "build" directory into "C:\mlpack\armadillo-8.500.1\"
+- Open the Command Prompt and navigate to "C:\mlpack\armadillo-8.500.1\build"
+- Run cmake: 
+
+@code
+cmake -G "Visual Studio 15 2017 Win64" -DBLAS_LIBRARY:FILEPATH="C:/mlpack/mlpack-3.0.2/packages/OpenBLAS.0.2.14.1/lib/native/lib/x64/libopenblas.dll.a" -DLAPACK_LIBRARY:FILEPATH="C:/mlpack/mlpack-3.0.2/packages/OpenBLAS.0.2.14.1/lib/native/lib/x64/libopenblas.dll.a" -DCMAKE_PREFIX:FILEPATH="C:/mlpack/armadillo" -DBUILD_SHARED_LIBS=OFF ..
+@endcode
+
[…]
