### 2018-06-13 7d4055b
initialize members in constructor
