+++ src/mlpack/methods/cf/normalization/overall_mean_normalization.hpp
@@ -40,7 +40,7 @@ class OverallMeanNormalization
 {
  public:
   // Empty constructor.
-  OverallMeanNormalization(): mean(0) { }
+  OverallMeanNormalization() : mean(0) { }
 
   /**
    * Normalize the data by subtracting the mean of all existing ratings.
