+++ src/mlpack/methods/lmnn/lmnn_main.cpp
@@ -19,6 +19,8 @@
 
 #include "lmnn.hpp"
 
+#include <mlpack/core/optimizers/bigbatch_sgd/bigbatch_sgd.hpp>
+#include <mlpack/core/optimizers/sgd/sgd.hpp>
 #include <mlpack/core/optimizers/lbfgs/lbfgs.hpp>
 
 // Define parameters.
@@ -36,15 +38,39 @@ PROGRAM_INFO("Large Margin Nearest Neighbors (LMNN)",
     "), or alternatively as a separate matrix (specified with " +
     PRINT_PARAM_STRING("labels") + ")."
     "\n\n"
-    "This implementation of LMNN uses stochastic gradient descent, mini-batch "
-    "stochastic gradient descent, or the L_BFGS optimizer. "
+    "This implementation of LMNN uses AdaGrad, BigBatch_SGD, stochastic "
+    "gradient descent, mini-batch stochastic gradient descent, or the L_BFGS "
+    "optimizer. "
+    "\n\n"
+    "AdaGrad, specified by the value 'adagrad' for the parameter " +
+    PRINT_PARAM_STRING("optimizer") + ", uses maximum of past squared "
+    "gradients. It primarily on six parameters: the step size (specified "
+    "with " + PRINT_PARAM_STRING("step_size") + "), the batch size (specified "
+    "with " + PRINT_PARAM_STRING("batch_size") + "), the maximum number of "
+    "passes (specified with " + PRINT_PARAM_STRING("passes") + ") "
+    ", the exponential decay rate for the first moment estimates (specified "
+    "with " + PRINT_PARAM_STRING("beta1") + "), the exponential decay rate for "
+    "weighted infinity norm (specified with " + PRINT_PARAM_STRING("beta2") +
+     "), epsilon (specified with " + PRINT_PARAM_STRING("epsilon") + ").  In "
+    "addition, a normalized starting point can be used by specifying the " +
+    PRINT_PARAM_STRING("normalize") + " parameter. "
+    "\n\n"
+    "BigBatch_SGD, specified by the value 'bbsgd' for the parameter " +
+    PRINT_PARAM_STRING("optimizer") + ", depends primarily on four parameters: "
+    "the step size (specified with " + PRINT_PARAM_STRING("step_size") + "), "
+    "the batch size (specified with " + PRINT_PARAM_STRING("batch_size") + "), "
+    "the maximum number of passes (specified with " +
+    PRINT_PARAM_STRING("passes") + "), the batch delta (specified "
+    "with " + PRINT_PARAM_STRING("batch_delta") + ").  In addition, "
+    "a normalized starting point can be used by specifying the " +
+    PRINT_PARAM_STRING("normalize") + " parameter. "
     "\n\n"
     "Stochastic gradient descent, specified by the value 'sgd' for the "
     "parameter " + PRINT_PARAM_STRING("optimizer") + ", depends "
     "primarily on three parameters: the step size (specified with " +
     PRINT_PARAM_STRING("step_size") + "), the batch size (specified with " +
-    PRINT_PARAM_STRING("batch_size") + "), and the maximum number of iterations"
-    " (specified with " + PRINT_PARAM_STRING("max_iterations") + ").  In "
+    PRINT_PARAM_STRING("batch_size") + "), and the maximum number of passes "
+    "(specified with " + PRINT_PARAM_STRING("passes") + ").  In "
     "addition, a normalized starting point can be used by specifying the " +
[…]
