+++ src/mlpack/methods/lmnn/lmnn_function_impl.hpp
@@ -31,8 +31,7 @@ LMNNFunction<MetricType>::LMNNFunction(const arma::mat& dataset,
     labels(math::MakeAlias(const_cast<arma::Row<size_t>&>(labels), false)),
     k(k),
     metric(metric),
-    regularization(regularization),
-    precalculated(false)
+    regularization(regularization)
 {
   // Initialize the initial learning point.
   initialPoint.eye(dataset.n_rows, dataset.n_rows);
@@ -177,15 +176,18 @@ void LMNNFunction<MetricType>::Gradient(const arma::mat& coordinates,
 {
   gradient.zeros(coordinates.n_rows, coordinates.n_cols);
 
-  // Calculate gradient due to target neighbors.
-  Precalculate(dataset.n_cols);
-  arma::mat cij = p_cij[0];
-
-  // Calculate gradient due to impostors.
+  arma::mat cij = arma::zeros(dataset.n_rows, dataset.n_rows);
   arma::mat cil = arma::zeros(dataset.n_rows, dataset.n_rows);
 
   for (size_t i = 0; i < dataset.n_cols; i++)
   {
+    for (size_t j = 0; j < k ; j++)
+    {
+      // Calculate gradient due to target neighbors.
+      arma::vec diff = dataset.col(i) - dataset.col(targetNeighbors(j, i));
+      cij += diff * arma::trans(diff);
+    }
+
     for (int j = k - 1; j >= 0; j--)
     {
       // Bound constraints to avoid uneccesary computation.
@@ -205,6 +207,7 @@ void LMNNFunction<MetricType>::Gradient(const arma::mat& coordinates,
           break;
         }
 
+        // Caculate gradient due to impostors.
         arma::vec diff = dataset.col(i) - dataset.col(targetNeighbors(j, i));
         cil += diff * arma::trans(diff);
 
@@ -228,15 +231,18 @@ void LMNNFunction<MetricType>::Gradient(const arma::mat& coordinates,
 {
   gradient.zeros(coordinates.n_rows, coordinates.n_cols);
 
-  // Calculate gradient due to target neighbors.
-  Precalculate(batchSize);
-  arma::mat cij = p_cij[begin / batchSize];
-
-  // Calculate gradient due to impostors.
+  arma::mat cij = arma::zeros(dataset.n_rows, dataset.n_rows);
   arma::mat cil = arma::zeros(dataset.n_rows, dataset.n_rows);
 
   for (size_t i = begin; i < begin + batchSize; i++)
   {
+    for (size_t j = 0; j < k ; j++)
+    {
+      // Calculate gradient due to target neighbors.
+      arma::vec diff = dataset.col(i) - dataset.col(targetNeighbors(j, i));
+      cij += diff * arma::trans(diff);
+    }
+
     for (int j = k - 1; j >= 0; j--)
     {
       // Bound constraints to avoid uneccesary computation.
@@ -256,6 +262,7 @@ void LMNNFunction<MetricType>::Gradient(const arma::mat& coordinates,
           break;
         }
 
+        // Caculate gradient due to impostors.
         arma::vec diff = dataset.col(i) - dataset.col(targetNeighbors(j, i));
         cil += diff * arma::trans(diff);
 
@@ -287,21 +294,21 @@ double LMNNFunction<MetricType>::EvaluateWithGradient(
 
   gradient.zeros(coordinates.n_rows, coordinates.n_cols);
 
-  // Calculate gradient due to target neighbors.
-  Precalculate(dataset.n_cols);
-  arma::mat cij = p_cij[0];
-
-  // Calculate gradient due to impostors.
+  arma::mat cij = arma::zeros(dataset.n_rows, dataset.n_rows);
[…]
