### 2018-06-13 059f3b1
Use precalculate for optimizers which do not uses batches
