### 2018-06-13 8a6709f
Added a range parameter to calculate impostors after some particular iterations
