### 2018-06-13 e357e8f
includeKl boolean added, tests added
