+++ src/mlpack/core/optimizers/dual_optimizer/CMakeLists.txt
@@ -1,0 +1,11 @@
+set(SOURCES
+  dual_optimizer.hpp
+  dual_optimizer_impl.hpp
+)
+
+set(DIR_SRCS)
+foreach(file ${SOURCES})
+  set(DIR_SRCS ${DIR_SRCS} ${CMAKE_CURRENT_SOURCE_DIR}/${file})
+endforeach()
+
+set(MLPACK_SRCS ${MLPACK_SRCS} ${DIR_SRCS} PARENT_SCOPE)
