+++ src/mlpack/methods/cf/cf.hpp
@@ -60,9 +60,6 @@ namespace cf /** Collaborative filtering. **/ {
  * @tparam NormalizationType The type of normalization performed on raw data.
  *     Data is normalized before calling Train() method. Predicted rating is
  *     denormalized before return.
- * 
- * @tparam DecompositionPolicy The algorithm to use to decompose
- *     the rating matrix (a W and H matrix).
  */
 template<typename NormalizationType = NoNormalization>
 class CFType
@@ -72,8 +69,7 @@ class CFType
    * Initialize the CFType object without performing any factorization.  Be sure to
    * call Train() before calling GetRecommendations() or any other functions!
    */
-  CFType(const size_t numUsersForSimilarity = 5,
-     const size_t rank = 0);
+  CFType(const size_t numUsersForSimilarity = 5, const size_t rank = 0);
 
   /**
    * Initialize the CFType object using any decomposition method, immediately
@@ -85,6 +81,12 @@ class CFType
    * The provided dataset can be a coordinate list; that is, a 3-row matrix
    * where each column corresponds to a (user, item, rating) entry in the
    * matrix or a sparse matrix representing (user, item) table.
+   *
+   * @tparam MatType The type of input matrix, which is expected to be either
+   *     arma::mat (table of (user, item, rating)) or arma::sp_mat (sparse
+   *     rating matrix where row is item and column is user).
+   * @tparam DecompositionPolicy The algorithm to use to decompose
+   *     the rating matrix (a W and H matrix).
    *
    * @param data Data matrix: dense matrix (coordinate lists) 
    *    or sparse matrix(cleaned).
@@ -109,6 +111,9 @@ class CFType
    * parameters that have already been set for the model (specifically, the rank
    * parameter), and optionally, using the given DecompositionPolicy.
    *
+   * @tparam DecompositionPolicy The algorithm to use to decompose
+   *     the rating matrix (a W and H matrix).
+   *
    * @param data Input dataset; dense matrix (coordinate lists).
    * @param decomposition Instantiated DecompositionPolicy object.
    * @param maxIterations Maximum number of iterations.
@@ -126,6 +131,9 @@ class CFType
    * Train the CFType model (i.e. factorize the input matrix) using the
    * parameters that have already been set for the model (specifically, the
    * rank parameter), and optionally, using the given DecompositionPolicy.
+   *
+   * @tparam DecompositionPolicy The algorithm to use to decompose
+   *     the rating matrix (a W and H matrix).
    *
    * @param data Input dataset; sparse matrix (user item table).
    * @param decomposition Instantiated DecompositionPolicy object.
