### 2018-06-14 5fbe03b
Denormalize(users(i), ...)
