### 2018-06-14 2554f60
change from arma::vec userMean to arma::rowvec userMean
