### 2018-06-14 #726 @rcurtin
Hello reviewers, thank you so much for your comments.  We've released mlpack 3.0.2 last week that addresses all the issues.  Is there anything else needed from our end?

Thanks!
