### 2018-06-14 #726 @Kevin-Mattheus-Moerman
Great. It looks like we are all set! @mlgill @MikeInnes @rasbt thanks for reviewing :tada:

@rcurtin please can you provide the DOI to an archived version of the reviewed software? You can use a service like [Zenodo](https://zenodo.org/), let me know if you have questions (see also https://guides.github.com/activities/citable-code/).
