### 2018-06-14 #726 @Kevin-Mattheus-Moerman
@arfon can you increment the version here to 3.0.2 and also proceed to accept once @rcurtin provides the DOI for the archived version? Thanks.
