### 2018-06-14 e08e761
Minor style fixes and add serialization.
