### 2018-06-15 35a97f4
normal dist added, reconstruction loss added
