### 2018-06-16 a011e7e
removed reduntant data members, moved NegativeLogLikelihood to loss folder
