+++ src/mlpack/methods/ann/layer/bilinear_interpolation.hpp
@@ -82,11 +82,6 @@ class BilinearInterpolation
                 arma::Mat<eT>&& gradient,
                 arma::Mat<eT>&& output);
 
-  //! Get the input parameter.
-  InputDataType const& InputParameter() const { return inputParameter; }
-  //! Modify the input parameter.
-  InputDataType& InputParameter() { return inputParameter; }
-
   //! Get the output parameter.
   OutputDataType const& OutputParameter() const { return outputParameter; }
   //! Modify the output parameter.
@@ -116,8 +111,6 @@ class BilinearInterpolation
   size_t depth;
   //! Locally-stored delta object.
   OutputDataType delta;
-  //! Locally-stored input parameter object.
-  InputDataType inputParameter;
   //! Locally-stored output parameter object.
   OutputDataType outputParameter;
 }; // class BilinearInterpolation
