+++ src/mlpack/methods/ann/layer/negative_log_likelihood.hpp
@@ -1,106 +1,0 @@
-/**
- * @file negative_log_likelihood.hpp
- * @author Marcus Edel
- *
- * Definition of the NegativeLogLikelihood class.
- *
- * mlpack is free software; you may redistribute it and/or modify it under the
- * terms of the 3-clause BSD license.  You should have received a copy of the
- * 3-clause BSD license along with mlpack.  If not, see
- * http://www.opensource.org/licenses/BSD-3-Clause for more information.
- */
-#ifndef MLPACK_METHODS_ANN_LAYER_NEGATIVE_LOG_LIKELIHOOD_HPP
-#define MLPACK_METHODS_ANN_LAYER_NEGATIVE_LOG_LIKELIHOOD_HPP
-
-#include <mlpack/prereqs.hpp>
-
-namespace mlpack {
-namespace ann /** Artificial Neural Network. */ {
-
-/**
- * Implementation of the negative log likelihood layer. The negative log
- * likelihood layer expectes that the input contains log-probabilities for each
- * class. The layer also expects a class index, in the range between 1 and the
- * number of classes, as target when calling the Forward function.
- *
- * @tparam InputDataType Type of the input data (arma::colvec, arma::mat,
- *         arma::sp_mat or arma::cube).
- * @tparam OutputDataType Type of the output data (arma::colvec, arma::mat,
- *         arma::sp_mat or arma::cube).
- */
-template <
-    typename InputDataType = arma::mat,
-    typename OutputDataType = arma::mat
->
-class NegativeLogLikelihood
-{
- public:
-  /**
-   * Create the NegativeLogLikelihoodLayer object.
-   */
-  NegativeLogLikelihood();
-
-  /*
-   * Computes the Negative log likelihood.
-   *
-   * @param input Input data used for evaluating the specified function.
-   * @param output Resulting output activation.
-   */
-  template<typename InputType, typename TargetType>
-  double Forward(const InputType&& input, TargetType&& target);
-
-  /**
-   * Ordinary feed backward pass of a neural network. The negative log
-   * likelihood layer expectes that the input contains log-probabilities for
-   * each class. The layer also expects a class index, in the range between 1
-   * and the number of classes, as target when calling the Forward function.
-   *
-   * @param input The propagated input activation.
-   * @param target The target vector, that contains the class index in the range
-   *        between 1 and the number of classes.
-   * @param output The calculated error.
-   */
-  template<typename InputType, typename TargetType, typename OutputType>
-  void Backward(const InputType&& input,
-                const TargetType&& target,
-                OutputType&& output);
-
-  //! Get the input parameter.
-  InputDataType& InputParameter() const { return inputParameter; }
-  //! Modify the input parameter.
-  InputDataType& InputParameter() { return inputParameter; }
-
-  //! Get the output parameter.
-  OutputDataType& OutputParameter() const { return outputParameter; }
-  //! Modify the output parameter.
-  OutputDataType& OutputParameter() { return outputParameter; }
-
-  //! Get the delta.
-  OutputDataType& Delta() const { return delta; }
-  //! Modify the delta.
-  OutputDataType& Delta() { return delta; }
-
-  /**
-   * Serialize the layer
-   */
-  template<typename Archive>
-  void serialize(Archive& /* ar */, const unsigned int /* version */);
-
- private:
-  //! Locally-stored delta object.
-  OutputDataType delta;
-
-  //! Locally-stored input parameter object.
-  InputDataType inputParameter;
-
-  //! Locally-stored output parameter object.
-  OutputDataType outputParameter;
-}; // class NegativeLogLikelihood
-
-} // namespace ann
-} // namespace mlpack
-
-// Include implementation.
-#include "negative_log_likelihood_impl.hpp"
-
-#endif
