+++ src/mlpack/methods/ann/layer/log_softmax.hpp
@@ -65,11 +65,6 @@ class LogSoftMax
                 arma::Mat<eT>&& gy,
                 arma::Mat<eT>&& g);
 
-  //! Get the input parameter.
-  InputDataType& InputParameter() const { return inputParameter; }
-  //! Modify the input parameter.
-  InputDataType& InputParameter() { return inputParameter; }
-
   //! Get the output parameter.
   OutputDataType& OutputParameter() const { return outputParameter; }
   //! Modify the output parameter.
@@ -89,9 +84,6 @@ class LogSoftMax
  private:
   //! Locally-stored delta object.
   OutputDataType delta;
-
-  //! Locally-stored input parameter object.
-  InputDataType inputParameter;
 
   //! Locally-stored output parameter object.
   OutputDataType outputParameter;
