+++ src/mlpack/methods/ann/layer/concat_performance.hpp
@@ -72,11 +72,6 @@ class ConcatPerformance
                 const arma::Mat<eT>&& target,
                 arma::Mat<eT>&& output);
 
-  //! Get the input parameter.
-  InputDataType& InputParameter() const { return inputParameter; }
-  //! Modify the input parameter.
-  InputDataType& InputParameter() { return inputParameter; }
-
   //! Get the output parameter.
   OutputDataType& OutputParameter() const { return outputParameter; }
   //! Modify the output parameter.
@@ -102,9 +97,6 @@ class ConcatPerformance
 
   //! Locally-stored delta object.
   OutputDataType delta;
-
-  //! Locally-stored input parameter object.
-  InputDataType inputParameter;
 
   //! Locally-stored output parameter object.
   OutputDataType outputParameter;
