### 2018-06-16 fa3481c
moved NegativeLogLikelihood to loss folder
