+++ src/mlpack/methods/ann/layer/sequential.hpp
@@ -124,9 +124,9 @@ class Sequential
   //! Modify the initial point for the optimization.
   arma::mat& Parameters() { return parameters; }
 
-  //! Get the output parameter.
+  //! Get the input parameter.
   arma::mat const& InputParameter() const { return inputParameter; }
-  //! Modify the output parameter.
+  //! Modify the input parameter.
   arma::mat& InputParameter() { return inputParameter; }
 
   //! Get the output parameter.
