### 2018-06-16 d04bf9b
support for matrix subview added
