### 2018-06-16 ce7628f
Update citation and comment style.
