+++ src/mlpack/core/optimizers/adam/nadamax_update.hpp
@@ -138,28 +138,28 @@ class NadaMaxUpdate
   double& ScheduleDecay() { return scheduleDecay; }
 
  private:
-  // The epsilon value used to initialise the squared gradient parameter.
+  //! The epsilon value used to initialise the squared gradient parameter.
   double epsilon;
 
-  // The smoothing parameter.
+  //! The smoothing parameter.
   double beta1;
 
-  // The second moment coefficient.
+  //! The second moment coefficient.
   double beta2;
 
-  // The exponential moving average of gradient values.
+  //! The exponential moving average of gradient values.
   arma::mat m;
 
-  // The exponentially weighted infinity norm.
+  //! The exponentially weighted infinity norm.
   arma::mat u;
 
-  // The decay parameter for decay coefficients
+  //! The decay parameter for decay coefficients
   double scheduleDecay;
 
-  // The cumulative product of decay coefficients
+  //! The cumulative product of decay coefficients
   double cumBeta1;
 
-  // The number of iterations.
+  //! The number of iterations.
   double iteration;
 };
 
