### 2018-06-17 8e548c2
advanced ctor
