### 2018-06-17 d3ade9a
avoid reloading/regenerating datasets in cf_test
