+++ src/mlpack/tests/cf_test.cpp
@@ -18,12 +18,18 @@
 #include <mlpack/methods/cf/decomposition_policies/regularized_svd_method.hpp>
 #include <mlpack/methods/cf/decomposition_policies/svd_complete_method.hpp>
 #include <mlpack/methods/cf/decomposition_policies/svd_incomplete_method.hpp>
-#include  <mlpack/methods/cf/normalization/no_normalization.hpp>
-#include  <mlpack/methods/cf/normalization/overall_mean_normalization.hpp>
-#include  <mlpack/methods/cf/normalization/user_mean_normalization.hpp>
-#include  <mlpack/methods/cf/normalization/item_mean_normalization.hpp>
-#include  <mlpack/methods/cf/normalization/z_score_normalization.hpp>
-#include  <mlpack/methods/cf/normalization/combined_normalization.hpp>
+#include <mlpack/methods/cf/normalization/no_normalization.hpp>
+#include <mlpack/methods/cf/normalization/overall_mean_normalization.hpp>
+#include <mlpack/methods/cf/normalization/user_mean_normalization.hpp>
+#include <mlpack/methods/cf/normalization/item_mean_normalization.hpp>
+#include <mlpack/methods/cf/normalization/z_score_normalization.hpp>
+#include <mlpack/methods/cf/normalization/combined_normalization.hpp>
+#include <mlpack/methods/cf/neighbor_search_policies/lmetric_search.hpp>
+#include <mlpack/methods/cf/neighbor_search_policies/cosine_search.hpp>
+#include <mlpack/methods/cf/neighbor_search_policies/pearson_search.hpp>
+#include <mlpack/methods/cf/interpolation_policies/average_interpolation.hpp>
+#include <mlpack/methods/cf/interpolation_policies/similarity_interpolation.hpp>
+#include <mlpack/methods/cf/interpolation_policies/regression_interpolation.hpp>
 
 #include <iostream>
 
@@ -36,6 +42,66 @@ BOOST_AUTO_TEST_SUITE(CFTest);
 using namespace mlpack;
 using namespace mlpack::cf;
 using namespace std;
+
+// Get train and test dataset. We save datasets as static variables
+// to avoid reloading and regenerating datasets.
+static void GetDatasets(arma::mat& dataset, arma::mat& savedCols)
+{
+  // Whether datasets have been initialized.
+  static bool datasetsInitialized = false;
+
+  static arma::mat initDataset;
+  static arma::mat initSavedCols(3, 50);
+
+  // If datasets have been initialized, just pass them to parameters.
+  if (datasetsInitialized)
+  {
+    dataset = initDataset;
+    savedCols = initSavedCols;
+    return;
+  }
+
+  // If datasets haven't been initialized, initialize them.
+  data::Load("GroupLensSmall.csv", initDataset);
+
+  // Save the columns we've removed.
+  initSavedCols.fill(/* random very large value */ 10000000);
+  size_t currentCol = 0;
+  for (size_t i = 0; i < initDataset.n_cols; ++i)
+  {
+    if (currentCol == 50)
+      break;
+
+    if (initDataset(2, i) > 4.5) // 5-star rating.
+    {
+      // Make sure we don't have this user yet.  This is a slow way to do this
+      // but I don't particularly care here because it's in the tests.
+      bool found = false;
+      for (size_t j = 0; j < currentCol; ++j)
+      {
+        if (initSavedCols(0, j) == initDataset(0, i))
+        {
+          found = true;
+          break;
+        }
+      }
+
+      // If this user doesn't already exist in initSavedCols, add them.
+      // Otherwise ignore this point.
+      if (!found)
+      {
+        initSavedCols.col(currentCol) = initDataset.col(i);
[…]
