### 2018-06-17 d48f4d3
templatize Predict() & GetRecommendations()
