+++ src/mlpack/methods/cf/cf_impl.hpp
@@ -141,6 +141,7 @@ void CFType<NormalizationType>::Train(const arma::sp_mat& data,
 }
 
 template<typename NormalizationType>
+template<typename NeighborSearchPolicy, typename InterpolationPolicy>
 void CFType<NormalizationType>::GetRecommendations(
     const size_t numRecs,
     arma::Mat<size_t>& recommendations)
@@ -153,10 +154,12 @@ void CFType<NormalizationType>::GetRecommendations(
       cleanedData.n_cols - 1, cleanedData.n_cols);
 
   // Call the main overload for recommendations.
-  GetRecommendations(numRecs, recommendations, users);
+  GetRecommendations<NeighborSearchPolicy,
+                     InterpolationPolicy>(numRecs, recommendations, users);
 }
 
 template<typename NormalizationType>
+template<typename NeighborSearchPolicy, typename InterpolationPolicy>
 void CFType<NormalizationType>::GetRecommendations(
     const size_t numRecs,
     arma::Mat<size_t>& recommendations,
@@ -186,30 +189,41 @@ void CFType<NormalizationType>::GetRecommendations(
   arma::Mat<size_t> neighborhood;
 
   // Calculate the neighborhood of the queried users.  Note that the query user
-  // is part of the neighborhood---this is intentional.  We want to use an
-  // average of both the query user and the local neighborhood of the query
-  // user.
-  // The neighbor search technique should be a template parameter.
-  neighbor::KNN a(stretchedH);
-  arma::mat resultingDistances; // Temporary storage.
-  a.Search(query, numUsersForSimilarity, neighborhood, resultingDistances);
+  // is part of the neighborhood---this is intentional.  We want to use the
+  // weighted sum of both the query user and the local neighborhood of the 
+  // query user.
+  // Calculate the neighborhood of the queried users.
+  NeighborSearchPolicy neighborSearch(stretchedH);
+  arma::mat similarities; // Resulting similarities.
+
+  neighborSearch.Search(
+      query, numUsersForSimilarity, neighborhood, similarities);
 
   // Generate recommendations for each query user by finding the maximum numRecs
-  // elements in the averages matrix.
+  // elements in the ratings vector.
   recommendations.set_size(numRecs, users.n_elem);
   arma::mat values(numRecs, users.n_elem);
   recommendations.fill(SIZE_MAX);
   values.fill(DBL_MAX);
 
+  // Initialization of an InterpolationPolicy object should be put ahead of the
+  // following loop, because the initialization may takes a relatively long
+  // time and we don't want to repeat the initialization process in each loop.
+  InterpolationPolicy interpolation(cleanedData);
+
   for (size_t i = 0; i < users.n_elem; i++)
   {
-    // First, calculate average of neighborhood values.
-    arma::vec averages;
-    averages.zeros(cleanedData.n_rows);
+    // First, calculate the weighted sum of neighborhood values.
+    arma::vec ratings;
+    ratings.zeros(cleanedData.n_rows);
+
+    // Calculate interpolation weights.
+    arma::vec weights;
+    interpolation.GetWeights(weights, w, h, users(i),
+        neighborhood.col(i), similarities.col(i), cleanedData);
 
     for (size_t j = 0; j < neighborhood.n_rows; ++j)
-      averages += w * h.col(neighborhood(j, i));
-    averages /= neighborhood.n_rows;
+      ratings += weights(j) * (w * h.col(neighborhood(j, i)));
 
     // Let's build the list of candidate recomendations for the given user.
     // Default candidate: the smallest possible value and invalid item number.
[…]
