### 2018-06-17 c8d57fa
add tests for neighbor search policies and interpolation policies
