### 2018-06-18 bf4446a
Add SWATS optimizer implementation.
