+++ src/mlpack/core/optimizers/swats/CMakeLists.txt
@@ -1,0 +1,12 @@
+set(SOURCES
+  swats.hpp
+  swats_impl.hpp
+  swats_update.hpp
+)
+
+set(DIR_SRCS)
+foreach(file ${SOURCES})
+  set(DIR_SRCS ${DIR_SRCS} ${CMAKE_CURRENT_SOURCE_DIR}/${file})
+endforeach()
+
+set(MLPACK_SRCS ${MLPACK_SRCS} ${DIR_SRCS} PARENT_SCOPE)
