### 2018-06-18 5620023
Add SWATS optimizer test suite.
