### 2018-06-18 #726 @rcurtin
I've posted the library to Zenodo: https://zenodo.org/record/1292120.  Let me know if anything is needed from the mlpack end.  Thank you so much!
