### 2018-06-18 mlpack/mlpack#1403 @rcurtin
I think this is fixed now so I will go ahead and close the issue.
