### 2018-06-18 #726 @Kevin-Mattheus-Moerman
@arfon we are good to accept here
